[water] reorganize index expr related files, NFC
Group all logic pertaining to index expressions in
IndexExprInference.h/cpp. This fixes the layering between interfaces
that need be op-agnostic and analysis using the interface that isn't.

Move the index expression inference pass into its own file from
InferTypes.cpp where it was originally bootstrapped to reuse lattice
utilities.
