+++ water/include/water/Dialect/Wave/IR/WaveInterfaces.h
@@ -786,13 +786,7 @@ private:
   constexpr static unsigned kUndecidableState = 2;
 };
 
-void operator<<(mlir::Diagnostic &diag, const IndexExprsLatticeStorage &value);
 } // namespace wave
-
-namespace llvm {
-llvm::raw_ostream &operator<<(llvm::raw_ostream &os,
-                              const wave::IndexExprsLatticeStorage &value);
-} // namespace llvm
 
 namespace wave {
 namespace detail {
@@ -807,49 +801,6 @@ identityIndexExprsPropagate(llvm::ArrayRef<IndexExprsLatticeStorage> from,
                             llvm::StringRef toName,
                             wave::EmitErrorFn emitError);
 
-// Heuristic to stop index expression propagation. It stops propagation of index
-// expressions based on the vector shape of the operation from which the index
-// expression originally propagates, stored in `sourceVectorShape` field of the
-// lattice. Propagation is skipped if:
-//
-//   1. Propagating towards an Mma operation.
-//   2. The source vector shape doesn't cover all symbolic dimensions of the
-//      value the lattice is about to be propagated to.
-//   3. Any of the symbols already present in the index expression for the
-//      target value (which is normally initialized from the target's value
-//      shape) corresponds to a unit dimension in the source vector shape IF the
-//      rank is less than the number of such non-unit dimensions.
-//   4. Any of the non-unit dimensions in the source vector shape is not already
-//      included in the target value's index expression IF the rank is greater
-//      than or equal to the number of such dimensions.
-//
-// XXX: conditions 2-4 are carried over from the python prototype and are not
-// principled.
-//
-// Defined in WaveOps.cpp to uss specific op names, which we don't want in
-// interfaces.
-// TODO: move all the index expr logic to one file and avoid this spreadout.
-bool shouldPropagateIndexExprs(const wave::IndexExprsLatticeStorage &from,
-                               const wave::IndexExprsLatticeStorage &to,
-                               mlir::Value toValue);
-
-// Build thread-independent index mapping for a single tensor type and append to
-// symbolMappings. Used by identity and reduction index expr initialization.
-// Defined in WaveOps.cpp to use mixInThreadIndependentConstraints function
-// template that needs access to specific ops, which we don't want in
-// interfaces.
-// TODO: move all the index expr logic to one file and avoid this spreadout.
-llvm::LogicalResult buildThreadIndependentIndexMappings(
-    mlir::Operation *op, mlir::Type type,
-    const IndexExprsAnalysisInit &initObject,
-    llvm::SmallVectorImpl<mlir::NamedAttribute> &symbolMappings);
-
-// Create a new vector shape dictionary attribute with only the provided symbols
-// present.
-mlir::DictionaryAttr
-filterVectorShape(mlir::DictionaryAttr vectorShape,
-                  llvm::ArrayRef<wave::WaveSymbolAttr> symbols);
-
 // Default implementation for interface: initialize index expressions with
 // thread-independent constraints for all values returned by
 // getIndexExprValuesAndDescriptions. Forward analysis initializes results,
@@ -863,15 +814,6 @@ llvm::LogicalResult defaultInitializeIndexExprsBackward(
     llvm::MutableArrayRef<IndexExprsLatticeStorage> operandExprs,
     const IndexExprsAnalysisInit &initObject, wave::EmitErrorFn emitError,
     wave::EmitDelayedErrorFn &delayedErrorEmitter);
-
-// Check the index expressions is a concrete value rather lattice top/bottom and
-// append it to the indexExprs list. If it is lattice top/bottom, report an
-// error and return failure.
-llvm::LogicalResult
-checkAndAppendIndexExpr(mlir::Location loc,
-                        const IndexExprsLatticeStorage &expr,
-                        const llvm::Twine &description,
-                        llvm::SmallVectorImpl<mlir::Attribute> &indexExprs);
 
 static inline std::function<void(llvm::raw_ostream &, unsigned)>
 defaultGetIndexExprValuesAndDescriptions(
