+++ water/lib/Dialect/Wave/IR/WaveInterfaces.cpp
@@ -574,146 +574,6 @@ bool wave::detail::isReductionTypeInferenceComplete(Value input, Value init,
       });
 }
 
-FailureOr<ChangeResult> wave::detail::propagateReductionIndexExprsForward(
-    TypeRange operandTypes, Value result,
-    llvm::ArrayRef<IndexExprsLatticeStorage> operandExprs,
-    llvm::MutableArrayRef<IndexExprsLatticeStorage> resultExprs,
-    EmitErrorFn emitError) {
-  auto targetTensorType = dyn_cast<WaveTensorType>(result.getType());
-  if (!targetTensorType) {
-    emitError() << "expected result tensor type, got " << result.getType();
-    return failure();
-  }
-
-  // Multiple operands appear after expansion, which requires index inference to
-  // work, so this is not expected to happen with correct pass pipeline setup.
-  if (operandExprs.size() != 2) {
-    emitError()
-        << "index inference not supported for reduction with multiple operands";
-    return failure();
-  }
-
-  // Forward propagation is identity only for symbols that are present.
-  return identityIndexExprsPropagate(
-             operandExprs[0].keepOnlySymbols(targetTensorType.getShape()),
-             resultExprs, result, "reduced value", "result", emitError) |
-         identityIndexExprsPropagate(operandExprs[1], resultExprs, result,
-                                     "init", "result", emitError);
-}
-
-FailureOr<ChangeResult> wave::detail::propagateReductionIndexExprsBackward(
-    ValueRange operands,
-    llvm::MutableArrayRef<IndexExprsLatticeStorage> operandExprs,
-    llvm::ArrayRef<IndexExprsLatticeStorage> resultExprs,
-    EmitErrorFn emitError) {
-  auto initTensorType = dyn_cast<WaveTensorType>(operands[1].getType());
-  if (!initTensorType) {
-    emitError() << "expected init tensor type, got " << operands[1].getType();
-    return failure();
-  }
-  if (operandExprs.size() != 2) {
-    emitError()
-        << "index inference not supported for reduction with multiple operands";
-    return failure();
-  }
-
-  // Backward propagation to the reduced is identity: it will propagate
-  // expressions for symbols present in both target and source, but additional
-  // propagation from the op defining the operand is needed to cover reduction
-  // dimensions. Backward propagation to the init is full identity.
-  return identityIndexExprsPropagate(resultExprs[0], operandExprs, operands,
-                                     "result", "operands", emitError) |
-         identityIndexExprsPropagate(
-             operandExprs[0].keepOnlySymbols(initTensorType.getShape()),
-             operandExprs[1], operands[1], "operand", "init", emitError) |
-         identityIndexExprsPropagate(operandExprs[1], operandExprs[0],
-                                     operands[0], "init", "operand", emitError);
-}
-
-//-----------------------------------------------------------------------------
-// Index expr initialization
-//-----------------------------------------------------------------------------
-
-mlir::DictionaryAttr
-wave::detail::filterVectorShape(mlir::DictionaryAttr vectorShape,
-                                llvm::ArrayRef<wave::WaveSymbolAttr> symbols) {
-  if (!vectorShape)
-    return vectorShape;
-  llvm::SmallVector<NamedAttribute> filtered =
-      llvm::filter_to_vector(vectorShape, [&](NamedAttribute attr) {
-        return llvm::find_if(symbols, [&](wave::WaveSymbolAttr symbol) {
-                 return symbol.getName() == attr.getName().getValue();
-               }) != symbols.end();
-      });
-  return DictionaryAttr::get(vectorShape.getContext(), filtered);
-}
-
-static void initializeIndexExprsWithThreadIndependentConstraints(
-    Operation *op, Type type, wave::IndexExprsLatticeStorage &storage,
-    const wave::IndexExprsAnalysisInit &initObject) {
-  llvm::SmallVector<mlir::NamedAttribute> symbolMappings;
-  if (failed(wave::detail::buildThreadIndependentIndexMappings(
-          op, type, initObject, symbolMappings)))
-    return;
-
-  auto tensorType = cast<wave::WaveTensorType>(type);
-  storage.unsafeSet(wave::IndexExprsLatticeStorage(
-      DictionaryAttr::get(op->getContext(), symbolMappings),
-      wave::IndexExprsLatticeStorage::kLowestPriority,
-      wave::detail::filterVectorShape(
-          initObject.hardwareConstraint.getVectorShapes(),
-          tensorType.getShape())));
-}
-
-LogicalResult wave::detail::defaultInitializeIndexExprsForward(
-    WaveInferIndexExprsOpInterface iface,
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> resultExprs,
-    const wave::IndexExprsAnalysisInit &initObject,
-    wave::EmitErrorFn emitError) {
-  SmallVector<Value> indexedValues;
-  iface.getIndexExprValuesAndDescriptions(indexedValues);
-  for (Value value : indexedValues) {
-    auto opResult = dyn_cast<OpResult>(value);
-    if (!opResult || opResult.getOwner() != iface.getOperation())
-      continue;
-
-    initializeIndexExprsWithThreadIndependentConstraints(
-        iface, opResult.getType(), resultExprs[opResult.getResultNumber()],
-        initObject);
-  }
-  return success();
-}
-
-LogicalResult wave::detail::defaultInitializeIndexExprsBackward(
-    WaveInferIndexExprsOpInterface iface,
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> operandExprs,
-    const wave::IndexExprsAnalysisInit &initObject, wave::EmitErrorFn emitError,
-    wave::EmitDelayedErrorFn &delayedErrorEmitter) {
-  SmallVector<Value> indexedValues;
-  iface.getIndexExprValuesAndDescriptions(indexedValues);
-  for (Value value : indexedValues) {
-    auto opResult = dyn_cast<OpResult>(value);
-    if (opResult && opResult.getOwner() == iface.getOperation())
-      continue;
-
-    // A value may be used repeatedly as operand, make sure to update all
-    // corresponding operand expression even though they go to the same lattice
-    // after all.
-    [[maybe_unused]] bool updated = false;
-    for (unsigned i = 0, e = operandExprs.size(); i < e; ++i) {
-      if (iface->getOperand(i) != value)
-        continue;
-
-      initializeIndexExprsWithThreadIndependentConstraints(
-          iface, iface->getOperand(i).getType(), operandExprs[i], initObject);
-      updated = true;
-    }
-    assert(updated && "value declared in getIndexExprValuesAndDescriptions is "
-                      "neither op operand nor result");
-  }
-  return success();
-}
-
 //-----------------------------------------------------------------------------
 // WaveElementsPerThreadOpInterface helpers
 //-----------------------------------------------------------------------------
@@ -1041,893 +901,6 @@ llvm::LogicalResult wave::detail::verifyCompatibleOperandsAndResultsOpTrait(
   return verifyTypeRange(op->getLoc(), op->getResultTypes(), referenceType,
                          includeAddressSpace, includeElementalType,
                          kResultNamePrefix, os.str());
-}
-
-//-----------------------------------------------------------------------------
-// Lattice implementation
-//-----------------------------------------------------------------------------
-
-wave::IndexExprsLatticeStorage::IndexExprsLatticeStorage()
-    : value(nullptr, kUninitializedState), vectorShape(nullptr) {}
-
-wave::IndexExprsLatticeStorage::IndexExprsLatticeStorage(
-    DictionaryAttr concreteValue, int32_t priority, DictionaryAttr vectorShape)
-    : value(concreteValue, kSpecificTypeState), vectorShape(vectorShape),
-      sourceVectorShape(priority == kLowestPriority ? nullptr : vectorShape),
-      sourceVectorShapePriority(priority) {
-  MLIRContext *ctx = concreteValue.getContext();
-  IntegerType i32 = IntegerType::get(ctx, 32);
-  IntegerAttr priAttr = IntegerAttr::get(i32, priority);
-  SmallVector<NamedAttribute> entries;
-  entries.reserve(concreteValue.size());
-  for (NamedAttribute attr : concreteValue)
-    entries.emplace_back(attr.getName(), priAttr);
-  priorities = DictionaryAttr::get(ctx, entries);
-}
-
-wave::IndexExprsLatticeStorage::IndexExprsLatticeStorage(
-    DictionaryAttr concreteValue, DictionaryAttr priorities,
-    DictionaryAttr vectorShape)
-    : value(concreteValue, kSpecificTypeState), priorities(priorities),
-      vectorShape(vectorShape), sourceVectorShape(vectorShape) {
-  if (priorities) {
-    for (NamedAttribute dimPriority : priorities)
-      sourceVectorShapePriority = std::max<int32_t>(
-          sourceVectorShapePriority,
-          llvm::cast<IntegerAttr>(dimPriority.getValue()).getInt());
-  }
-  if (sourceVectorShapePriority == kLowestPriority)
-    sourceVectorShape = nullptr;
-}
-
-wave::IndexExprsLatticeStorage::IndexExprsLatticeStorage(
-    DictionaryAttr concreteValue, DictionaryAttr priorities,
-    DictionaryAttr vectorShape, DictionaryAttr sourceVectorShape,
-    int32_t sourceVectorShapePriority)
-    : value(concreteValue, kSpecificTypeState), priorities(priorities),
-      vectorShape(vectorShape),
-      sourceVectorShape(sourceVectorShapePriority == kLowestPriority
-                            ? nullptr
-                            : sourceVectorShape),
-      sourceVectorShapePriority(sourceVectorShapePriority) {}
-
-bool wave::IndexExprsLatticeStorage::operator==(
-    const IndexExprsLatticeStorage &other) const {
-  return value == other.value && priorities == other.priorities &&
-         vectorShape == other.vectorShape &&
-         sourceVectorShape == other.sourceVectorShape &&
-         sourceVectorShapePriority == other.sourceVectorShapePriority;
-}
-
-bool wave::IndexExprsLatticeStorage::operator!=(
-    const IndexExprsLatticeStorage &other) const {
-  return !(*this == other);
-}
-
-bool wave::IndexExprsLatticeStorage::isBottom() const {
-  return value.getInt() == kUninitializedState;
-}
-
-bool wave::IndexExprsLatticeStorage::isTop() const {
-  return value.getInt() == kUndecidableState;
-}
-
-DictionaryAttr wave::IndexExprsLatticeStorage::getConcreteValue() const {
-  if (value.getInt() != kSpecificTypeState)
-    return nullptr;
-  return llvm::cast<DictionaryAttr>(value.getPointer());
-}
-
-DictionaryAttr wave::IndexExprsLatticeStorage::getVectorShape() const {
-  if (value.getInt() != kSpecificTypeState)
-    return nullptr;
-  return vectorShape;
-}
-
-DictionaryAttr wave::IndexExprsLatticeStorage::getSourceVectorShape() const {
-  if (value.getInt() != kSpecificTypeState)
-    return nullptr;
-  return sourceVectorShape;
-}
-
-int32_t wave::IndexExprsLatticeStorage::getSourceVectorShapePriority() const {
-  return sourceVectorShapePriority;
-}
-
-wave::IndexExprsLatticeStorage
-wave::IndexExprsLatticeStorage::withSourceVectorShape(DictionaryAttr shape,
-                                                      int32_t priority) const {
-  IndexExprsLatticeStorage copy = *this;
-  copy.sourceVectorShape = shape;
-  copy.sourceVectorShapePriority = priority;
-  return copy;
-}
-
-int32_t
-wave::IndexExprsLatticeStorage::getPriorityForKey(StringAttr key) const {
-  assert(getConcreteValue() && "no priority for lattice top/bottom");
-  if (!priorities)
-    return kLowestPriority;
-  Attribute val = priorities.get(key);
-  if (!val)
-    return kLowestPriority;
-  return llvm::cast<IntegerAttr>(val).getInt();
-}
-
-wave::IndexExprsLatticeStorage wave::IndexExprsLatticeStorage::top() {
-  IndexExprsLatticeStorage result;
-  result.value.setPointer(nullptr);
-  result.value.setInt(kUndecidableState);
-  result.vectorShape = nullptr;
-  result.sourceVectorShape = nullptr;
-  result.sourceVectorShapePriority = 0;
-  return result;
-}
-
-wave::IndexExprsLatticeStorage wave::IndexExprsLatticeStorage::bottom() {
-  IndexExprsLatticeStorage result;
-  result.value.setPointer(nullptr);
-  result.value.setInt(kUninitializedState);
-  result.vectorShape = nullptr;
-  result.sourceVectorShape = nullptr;
-  result.sourceVectorShapePriority = 0;
-  return result;
-}
-
-/// Parse and validate wave constraints from an attribute array.
-/// Returns the hardware constraint or nullptr on failure.
-static wave::HardwareConstraintAttr parseWaveConstraints(
-    Location loc, Attribute constraints,
-    llvm::DenseMap<wave::WaveSymbolAttr, llvm::SmallVector<Attribute>>
-        &symbolConstraints) {
-  wave::HardwareConstraintAttr hardwareConstraint;
-  for (Attribute constraint : llvm::cast<ArrayAttr>(constraints)) {
-    if (auto workgroup =
-            llvm::dyn_cast<wave::WorkgroupConstraintAttr>(constraint)) {
-      symbolConstraints[workgroup.getDim()].push_back(workgroup);
-    } else if (auto tiling =
-                   llvm::dyn_cast<wave::TilingConstraintAttr>(constraint)) {
-      symbolConstraints[tiling.getDim()].push_back(tiling);
-    } else if (auto waveConstraint =
-                   llvm::dyn_cast<wave::WaveConstraintAttr>(constraint)) {
-      symbolConstraints[waveConstraint.getDim()].push_back(waveConstraint);
-    } else if (auto hardware =
-                   llvm::dyn_cast<wave::HardwareConstraintAttr>(constraint)) {
-      assert(hardwareConstraint == nullptr &&
-             "multiple hardware constraints are not supported");
-      hardwareConstraint = hardware;
-    } else {
-      emitError(loc) << "unsupported constraint type: " << constraint;
-      return nullptr;
-    }
-  }
-
-  if (!hardwareConstraint) {
-    emitError(loc) << "expected a hardware constraint";
-    return nullptr;
-  }
-
-  return hardwareConstraint;
-}
-
-llvm::FailureOr<wave::IndexExprsAnalysisInit>
-wave::IndexExprsAnalysisInit::create(Operation *parent,
-                                     Attribute constraintsAttr,
-                                     wave::WaveHyperparameterAttr hyperparams) {
-  Location loc = parent->getLoc();
-  wave::IndexExprsAnalysisInit initObject;
-  initObject.hardwareConstraint =
-      parseWaveConstraints(loc, constraintsAttr, initObject.symbolConstraints);
-  if (initObject.hardwareConstraint == nullptr)
-    return llvm::failure();
-
-  parent->walk(
-      [&](Operation *op) { initObject.deterministicOpOrder.push_back(op); });
-
-  // If waves_per_block is explicitly provided, copy it to storage. Note that we
-  // have verified they match the result of dividing block tiles with wave tiles
-  // previously.
-  if (!initObject.hardwareConstraint.getWavesPerBlock().empty()) {
-    assert(initObject.hardwareConstraint.getWavesPerBlock().size() == 3 &&
-           "expected waves_per_block to have 3 elements");
-    llvm::ArrayRef<unsigned> explicitWpb =
-        initObject.hardwareConstraint.getWavesPerBlock();
-    initObject.wavesPerBlock.assign(explicitWpb);
-    return initObject;
-  }
-
-  // Otherwise, compute waves_per_block from wave constraints.
-  // First, extract wave and workgroup constraints from symbolConstraints.
-  llvm::SmallDenseMap<wave::WaveSymbolAttr, wave::WorkgroupConstraintAttr>
-      workgroupConstraints;
-  llvm::SmallDenseMap<wave::WaveSymbolAttr, wave::WaveConstraintAttr>
-      waveConstraints;
-
-  for (auto &&[symbol, constraints] : initObject.symbolConstraints) {
-    for (Attribute constraint : constraints) {
-      if (auto wg = llvm::dyn_cast<wave::WorkgroupConstraintAttr>(constraint))
-        workgroupConstraints[symbol] = wg;
-      else if (auto wv = llvm::dyn_cast<wave::WaveConstraintAttr>(constraint))
-        waveConstraints[symbol] = wv;
-    }
-  }
-
-  // If there are no wave constraints, default to [1, 1, 1].
-  if (waveConstraints.empty()) {
-    return emitError(loc) << "expected either waves_per_block in the hardware "
-                             "constraint or wave constraints on an ancestor op";
-  }
-
-  if (!hyperparams) {
-    return emitError(loc) << "cannot compute waves_per_block: missing "
-                             "hyperparameters attribute";
-  }
-
-  if (failed(wave::computeWavesPerBlockFromConstraints(
-          workgroupConstraints, waveConstraints, hyperparams,
-          initObject.wavesPerBlock))) {
-    return emitError(loc)
-           << "failed to compute waves_per_block from wave constraints";
-  }
-  return initObject;
-}
-
-// Create a new map with per-result sum between a and b maps.
-static AffineMap addMaps(AffineMap a, AffineMap b) {
-  assert(a.getNumResults() == b.getNumResults() &&
-         "expected both maps to have the same number of expressions");
-  SmallVector<AffineExpr> subtracted = llvm::map_to_vector(
-      llvm::zip_equal(a.getResults(), b.getResults()),
-      [&](auto &&pair) { return std::get<0>(pair) + std::get<1>(pair); });
-  return AffineMap::get(a.getNumDims(), a.getNumSymbols(), subtracted,
-                        a.getContext());
-}
-
-// Create a new map with per-result difference between a and b maps.
-static AffineMap subtractMaps(AffineMap a, AffineMap b) {
-  assert(a.getNumResults() == b.getNumResults() &&
-         "expected both maps to have the same number of expressions");
-  SmallVector<AffineExpr> subtracted = llvm::map_to_vector(
-      llvm::zip_equal(a.getResults(), b.getResults()),
-      [&](auto &&pair) { return std::get<0>(pair) - std::get<1>(pair); });
-  return AffineMap::get(a.getNumDims(), a.getNumSymbols(), subtracted,
-                        a.getContext());
-}
-
-const static wave::WaveIndexSymbol threadLikeIndexSymbols[] = {
-    wave::WaveIndexSymbol::THREAD_0, wave::WaveIndexSymbol::THREAD_1,
-    wave::WaveIndexSymbol::THREAD_2, wave::WaveIndexSymbol::GPR_NUMBER};
-
-// Return the list of thread-like index symbols.
-// TODO: It would be nice to cache the list somehow, but we need to tie it to
-// the context and ensure thread safety, potentially by storing it as an
-// attribute or some other named/typed entity in the context object.
-static SmallVector<Attribute> getThreadLikeIndexSymbols(MLIRContext *ctx) {
-  return llvm::map_to_vector(
-      threadLikeIndexSymbols, [&](wave::WaveIndexSymbol symbol) -> Attribute {
-        return wave::WaveIndexSymbolAttr::get(ctx, symbol);
-      });
-}
-
-// Return the list of index symbols other than thread-like.
-static SmallVector<Attribute> getNonThreadLikeIndexSymbols(MLIRContext *ctx) {
-  return llvm::map_to_vector(ArrayRef{wave::WaveIndexSymbol::WORKGROUP_0,
-                                      wave::WaveIndexSymbol::WORKGROUP_1,
-                                      wave::WaveIndexSymbol::WORKGROUP_2,
-                                      wave::WaveIndexSymbol::DEVICE_DIM_0,
-                                      wave::WaveIndexSymbol::DEVICE_DIM_1,
-                                      wave::WaveIndexSymbol::DEVICE_DIM_2},
-                             [&](wave::WaveIndexSymbol symbol) -> Attribute {
-                               return wave::WaveIndexSymbolAttr::get(ctx,
-                                                                     symbol);
-                             });
-}
-
-// Get the positions of `symbols` in `allSymbols`. Missing symbols are ignored.
-SmallVector<unsigned> getPositionsOfSymbols(ArrayRef<Attribute> symbols,
-                                            ArrayRef<Attribute> allSymbols) {
-  // Find positions of threadLikeIndexSymbols in symbols
-  SmallVector<unsigned> positions;
-  for (Attribute symbol : symbols) {
-    auto it = llvm::find(allSymbols, symbol);
-    if (it != allSymbols.end())
-      positions.push_back(std::distance(allSymbols.begin(), it));
-  }
-  return positions;
-}
-
-// Return true if any expression in the map is a function of a symbol at any of
-// the given positions.
-static bool isIndexExprMapFunctionOf(AffineMap map,
-                                     ArrayRef<unsigned> positions) {
-  return llvm::any_of(positions, [&](unsigned position) {
-    return map.isFunctionOfSymbol(position);
-  });
-}
-
-// Affine maps used in an index expression conceptually consists of multiple
-// additive components:
-//
-//   - thread independent component (workgroup and device indices)
-//   - thread dependent component (thread and GPR indices)
-//   - one component per iter dimension
-//
-// Two start maps can be joined if, for all pairwise components:
-//
-//   - the components are equal;
-//   - the component is absent from one of the maps.
-//
-// The join is then the sum of unique components from both maps.
-//
-// We check this by examining the difference between the two maps, which should
-// only contain symbols absent from one of the maps, i.e. symbols from the
-// symmetric difference of the symbol sets or, alternatively, not contain any
-// symbols from the intersection of the symbol sets. The difference should also
-// not contain a constant term.
-//
-// Additional care is taken for index (non-iter) dimensions as they must appear
-// or not appear simultaneously. For example, lhs may only have thread 0 index
-// and rhs may only have thread 1 index, so the difference will depend on both
-// thread 0 and thread 1 indices without either of them being common, so the
-// regular check won't detect that. Check for dependency on any such symbol
-// instead.
-//
-// The function takes the list of symbols used in LHS and RHS and separately a
-// list containing a union thereof and a list of positions in that list of the
-// common symbols. This is done for efficiency reasons to avoid re-computing
-// these several times when handling start/size/stride maps that share the
-// symbol lists.
-//
-// TODO: consider having separate expressions for each component for simplicity;
-// even further, consider having a lattice that is a collection of constraints
-// applicable to the value + metadata (like it being used in LHS/RHS/Acc of an
-// MMA) without creating the expression immediately.
-static FailureOr<AffineMap> getIndexExprStartJoinedMap(
-    AffineMap lhs, AffineMap rhs, ArrayRef<Attribute> lhsSymbols,
-    ArrayRef<Attribute> rhsSymbols, ArrayRef<Attribute> allSymbols,
-    ArrayRef<Attribute> commonSymbols) {
-  if (!lhs)
-    return rhs;
-  if (!rhs)
-    return lhs;
-
-  lhs = wave::alignMapSymbols(lhs, lhsSymbols, allSymbols);
-  rhs = wave::alignMapSymbols(rhs, rhsSymbols, allSymbols);
-
-  if (lhs == rhs)
-    return lhs;
-
-  AffineMap difference = simplifyAffineMap(subtractMaps(lhs, rhs));
-
-  MLIRContext *ctx = rhs.getContext();
-  SmallVector<unsigned> threadLikePositions =
-      getPositionsOfSymbols(getThreadLikeIndexSymbols(ctx), allSymbols);
-  SmallVector<unsigned> nonThreadLikePositions =
-      getPositionsOfSymbols(getNonThreadLikeIndexSymbols(ctx), allSymbols);
-  if (isIndexExprMapFunctionOf(difference, threadLikePositions) &&
-      isIndexExprMapFunctionOf(lhs, threadLikePositions) &&
-      isIndexExprMapFunctionOf(rhs, threadLikePositions)) {
-    return failure();
-  }
-  if (isIndexExprMapFunctionOf(difference, nonThreadLikePositions) &&
-      isIndexExprMapFunctionOf(lhs, nonThreadLikePositions) &&
-      isIndexExprMapFunctionOf(rhs, nonThreadLikePositions)) {
-    return failure();
-  }
-
-  // The symbolic part of the difference should not depend on any of the
-  // disallowed symbols (usually meaning symbols appearing in both).
-  for (AffineExpr expr : difference.getResults()) {
-    for (Attribute symbol : commonSymbols) {
-      auto it = llvm::find(allSymbols, symbol);
-      assert(it != allSymbols.end() &&
-             "expected common symbols to be a subset of all symbols");
-      unsigned position = std::distance(allSymbols.begin(), it);
-      if (expr.isFunctionOfSymbol(position) &&
-          !llvm::is_contained(threadLikePositions, position))
-        return failure();
-    }
-  }
-
-  // The constant parts of the expression must be equal.
-  // TODO: consider whether we want to allow one of the sides being 0 here. If
-  // we do, we will have to be more careful to construct a constant difference
-  // map here instead of taking the RHS constant in subtraction below.
-  AffineExpr zeroExpr = getAffineConstantExpr(0, ctx);
-  SmallVector<AffineExpr> symReplacements(allSymbols.size(), zeroExpr);
-  AffineMap lhsConstant =
-      lhs.replaceDimsAndSymbols(/*dimReplacements=*/{}, symReplacements,
-                                /*numResultDims=*/0, /*numResultSyms=*/0);
-  AffineMap rhsConstant =
-      rhs.replaceDimsAndSymbols(/*dimReplacements=*/{}, symReplacements,
-                                /*numResultDims=*/0, /*numResultSyms=*/0);
-  for (auto [lhsConstantExpr, rhsConstantExpr] :
-       llvm::zip_equal(lhsConstant.getResults(), rhsConstant.getResults())) {
-    auto lhsConstantExprCast = cast<AffineConstantExpr>(lhsConstantExpr);
-    auto rhsConstantExprCast = cast<AffineConstantExpr>(rhsConstantExpr);
-    if (lhsConstantExprCast.getValue() != rhsConstantExprCast.getValue())
-      return failure();
-  }
-
-  // Obtain a part of the RHS map that is only a function of RHS-specific
-  // symbols. For this, we replace all symbols appearing in the LHS map with
-  // zero. Symbol replacements contain zeros at this point. Reuse that but set
-  // RHS-only symbols to be replaced with themselves. Don't forget to subtract
-  // the constant part of RHS, which is known to be identical to that of RHS, so
-  // we don't duplicate it. At this point, we expect the caller to have removed
-  // unused symbols from the symbol list.
-  SmallVector<unsigned> lhsSymbolPositions =
-      getPositionsOfSymbols(lhsSymbols, allSymbols);
-  for (unsigned i = 0, e = symReplacements.size(); i < e; ++i) {
-    if (llvm::is_contained(lhsSymbolPositions, i))
-      continue;
-    symReplacements[i] = getAffineSymbolExpr(i, ctx);
-  }
-  AffineMap rhsOnly = rhs.replaceDimsAndSymbols(
-      {}, symReplacements, /*numResultDims=*/0, rhs.getNumSymbols());
-  rhsOnly = subtractMaps(rhsOnly, rhsConstant);
-  return simplifyAffineMap(addMaps(lhs, rhsOnly));
-}
-
-// Two step/stride maps can be joined if one of them is a constant one, at which
-// point the join is the other map, or if they are equal. All other combinations
-// join to lattice top.
-static FailureOr<AffineMap> getIndexExprStepStrideJoinedMap(
-    AffineMap lhs, AffineMap rhs, ArrayRef<Attribute> lhsSymbols,
-    ArrayRef<Attribute> rhsSymbols, ArrayRef<Attribute> allSymbols) {
-  if (!lhs)
-    return rhs;
-  if (!rhs)
-    return lhs;
-
-  lhs = wave::alignMapSymbols(lhs, lhsSymbols, allSymbols);
-  rhs = wave::alignMapSymbols(rhs, rhsSymbols, allSymbols);
-
-  if (lhs == rhs)
-    return lhs;
-
-  AffineExpr lhsExpr = lhs.getResult(0);
-  AffineExpr rhsExpr = rhs.getResult(0);
-  auto isConstantOne = [](AffineExpr expr) -> bool {
-    if (auto constantExpr = llvm::dyn_cast<AffineConstantExpr>(expr)) {
-      return constantExpr.getValue() == 1;
-    }
-    return false;
-  };
-  bool lhsIsConstantOne = isConstantOne(lhsExpr);
-  bool rhsIsConstantOne = isConstantOne(rhsExpr);
-  if (lhsIsConstantOne)
-    return rhs;
-  if (rhsIsConstantOne)
-    return lhs;
-  return failure();
-}
-
-// Join two concrete index expressions mappings either by picking the
-// higher-priority one or by joining their start/step/stride maps independently.
-// See getIndexExprStartJoinedMap and getIndexExprStepStrideJoinedMap for more
-// details on independent joining.
-static wave::WaveIndexMappingAttr
-getIndexExprsJoinMappings(wave::WaveIndexMappingAttr lhs,
-                          wave::WaveIndexMappingAttr rhs, int32_t lhsPriority,
-                          int32_t rhsPriority) {
-  if (lhsPriority > rhsPriority)
-    return lhs;
-  if (rhsPriority > lhsPriority)
-    return rhs;
-
-  // Collect all unique symbol names from both index mappings in order.
-  llvm::SmallVector<Attribute> allSymbols;
-  llvm::SetVector<Attribute> lhsSymbols(llvm::from_range, lhs.getSymbols());
-  llvm::SetVector<Attribute> rhsSymbols(llvm::from_range, rhs.getSymbols());
-  wave::aggregateAllSymbols(llvm::ArrayRef{lhsSymbols, rhsSymbols}, allSymbols);
-  llvm::SetVector<Attribute> commonSymbols =
-      llvm::set_intersection(lhsSymbols, rhsSymbols);
-
-  FailureOr<AffineMap> joinedStart = getIndexExprStartJoinedMap(
-      lhs.getStart(), rhs.getStart(), lhsSymbols.getArrayRef(),
-      rhsSymbols.getArrayRef(), allSymbols, commonSymbols.getArrayRef());
-  if (failed(joinedStart))
-    return nullptr;
-  FailureOr<AffineMap> joinedStep = getIndexExprStepStrideJoinedMap(
-      lhs.getStep(), rhs.getStep(), lhsSymbols.getArrayRef(),
-      rhsSymbols.getArrayRef(), allSymbols);
-  if (failed(joinedStep))
-    return nullptr;
-  FailureOr<AffineMap> joinedStride = getIndexExprStepStrideJoinedMap(
-      lhs.getStride(), rhs.getStride(), lhsSymbols.getArrayRef(),
-      rhsSymbols.getArrayRef(), allSymbols);
-  if (failed(joinedStride))
-    return nullptr;
-
-  return wave::WaveIndexMappingAttr::get(
-      lhs.getContext(), allSymbols, *joinedStart, *joinedStep, *joinedStride);
-}
-
-// Returns a joined vector shape using per-key priorities. For each key, the
-// higher-priority entry wins. If priorities are equal and values differ,
-// returns nullptr indicating the failure to join (reaching top).
-static DictionaryAttr getJoinedVectorShape(DictionaryAttr lhs,
-                                           DictionaryAttr rhs,
-                                           DictionaryAttr lhsPriorities,
-                                           DictionaryAttr rhsPriorities) {
-  if (!lhs)
-    return rhs;
-  if (!rhs)
-    return lhs;
-
-  if (lhs == rhs)
-    return lhs;
-
-  auto getPriority = [](DictionaryAttr map, StringAttr key) -> int32_t {
-    if (!map)
-      return wave::IndexExprsLatticeStorage::kLowestPriority;
-    Attribute val = map.get(key);
-    if (!val)
-      return wave::IndexExprsLatticeStorage::kLowestPriority;
-    return llvm::cast<IntegerAttr>(val).getInt();
-  };
-
-  SmallVector<NamedAttribute> joinedVectorShapeEntries;
-  joinedVectorShapeEntries.reserve(lhs.size() + rhs.size());
-
-  llvm::DenseSet<StringAttr> visited;
-  for (const NamedAttribute &attr : lhs) {
-    visited.insert(attr.getName());
-    Attribute rhsValue = rhs.get(attr.getName());
-    if (!rhsValue) {
-      joinedVectorShapeEntries.push_back(attr);
-      continue;
-    }
-    int32_t lhsPriority = getPriority(lhsPriorities, attr.getName());
-    int32_t rhsPriority = getPriority(rhsPriorities, attr.getName());
-    if (lhsPriority > rhsPriority) {
-      joinedVectorShapeEntries.push_back(attr);
-    } else if (rhsPriority > lhsPriority) {
-      joinedVectorShapeEntries.emplace_back(attr.getName(), rhsValue);
-    } else {
-      if (attr.getValue() != rhsValue)
-        return nullptr;
-      joinedVectorShapeEntries.push_back(attr);
-    }
-  }
-
-  for (const NamedAttribute &attr : rhs) {
-    if (!visited.contains(attr.getName()))
-      joinedVectorShapeEntries.push_back(attr);
-  }
-
-  return DictionaryAttr::get(lhs.getContext(), joinedVectorShapeEntries);
-}
-
-FailureOr<std::pair<DictionaryAttr, int32_t>>
-wave::IndexExprsLatticeStorage::getJoinedSourceVectorShape(
-    const IndexExprsLatticeStorage &lhs, const IndexExprsLatticeStorage &rhs) {
-  DictionaryAttr lhsSVS = lhs.getSourceVectorShape();
-  DictionaryAttr rhsSVS = rhs.getSourceVectorShape();
-  int32_t lhsPri = lhs.getSourceVectorShapePriority();
-  int32_t rhsPri = rhs.getSourceVectorShapePriority();
-
-  if (!lhsSVS && !rhsSVS)
-    return std::make_pair(DictionaryAttr(),
-                          wave::IndexExprsLatticeStorage::kLowestPriority);
-  if (!lhsSVS)
-    return std::make_pair(rhsSVS, rhsPri);
-  if (!rhsSVS)
-    return std::make_pair(lhsSVS, lhsPri);
-  if (lhsPri > rhsPri)
-    return std::make_pair(lhsSVS, lhsPri);
-  if (rhsPri > lhsPri)
-    return std::make_pair(rhsSVS, rhsPri);
-  if (lhsSVS == rhsSVS)
-    return std::make_pair(lhsSVS, lhsPri);
-  return failure();
-}
-
-FailureOr<DictionaryAttr> wave::IndexExprsLatticeStorage::getJoinedVectorShape(
-    const IndexExprsLatticeStorage &lhs, const IndexExprsLatticeStorage &rhs) {
-  if (lhs.isBottom() && rhs.isBottom())
-    return DictionaryAttr();
-  if (lhs.isTop() || rhs.isTop())
-    return failure();
-  if (lhs.isBottom())
-    return rhs.getVectorShape();
-  if (rhs.isBottom())
-    return lhs.getVectorShape();
-  if (DictionaryAttr result =
-          ::getJoinedVectorShape(lhs.getVectorShape(), rhs.getVectorShape(),
-                                 lhs.getPriorities(), rhs.getPriorities()))
-    return result;
-  return failure();
-}
-
-wave::IndexExprsLatticeStorage
-wave::IndexExprsLatticeStorage::join(const IndexExprsLatticeStorage &lhs,
-                                     const IndexExprsLatticeStorage &rhs) {
-  if (lhs == rhs)
-    return lhs;
-
-  // Top is saturating.
-  if (lhs.isTop() || rhs.isTop())
-    return top();
-
-  // Bottom is neutral.
-  if (lhs.isBottom())
-    return rhs;
-  if (rhs.isBottom())
-    return lhs;
-
-  MLIRContext *ctx = lhs.getConcreteValue().getContext();
-  DictionaryAttr lhsDict = lhs.getConcreteValue();
-  DictionaryAttr rhsDict = rhs.getConcreteValue();
-
-  // Join specific values per symbol using per-key priorities.
-  IntegerType i32 = IntegerType::get(ctx, 32);
-  llvm::DenseMap<StringAttr, Attribute> result;
-  llvm::DenseMap<StringAttr, int32_t> resultPriorities;
-  for (NamedAttribute namedAttr : lhsDict) {
-    result[namedAttr.getName()] = namedAttr.getValue();
-    resultPriorities[namedAttr.getName()] =
-        lhs.getPriorityForKey(namedAttr.getName());
-  }
-  for (NamedAttribute namedAttr : rhsDict) {
-    StringAttr key = namedAttr.getName();
-    int32_t rhsPriority = rhs.getPriorityForKey(key);
-
-    auto [it, inserted] = result.try_emplace(key, namedAttr.getValue());
-    if (inserted) {
-      resultPriorities[key] = rhsPriority;
-      continue;
-    }
-
-    int32_t lhsPriority = lhs.getPriorityForKey(key);
-
-    // The symbol has a mapping on both LHS and RHS, join them.
-    auto lhsMapping = llvm::cast<wave::WaveIndexMappingAttr>(it->getSecond());
-    auto rhsMapping =
-        llvm::cast<wave::WaveIndexMappingAttr>(namedAttr.getValue());
-    if (lhsMapping == rhsMapping) {
-      resultPriorities[key] = std::max(lhsPriority, rhsPriority);
-      continue;
-    }
-
-    wave::WaveIndexMappingAttr joinedMapping = getIndexExprsJoinMappings(
-        lhsMapping, rhsMapping, lhsPriority, rhsPriority);
-    if (!joinedMapping)
-      return IndexExprsLatticeStorage::top();
-
-    result[key] = joinedMapping;
-    resultPriorities[key] = std::max(lhsPriority, rhsPriority);
-  }
-
-  DictionaryAttr joinedVectorShape =
-      ::getJoinedVectorShape(lhs.getVectorShape(), rhs.getVectorShape(),
-                             lhs.getPriorities(), rhs.getPriorities());
-  if (!joinedVectorShape && (lhs.getVectorShape() || rhs.getVectorShape()))
-    return IndexExprsLatticeStorage::top();
-
-  // Join source vector shapes using priority: higher priority wins. If
-  // priorities are equal, two different values cause top; two identical values
-  // join cleanly.
-  auto joinedSVSResult = getJoinedSourceVectorShape(lhs, rhs);
-  if (failed(joinedSVSResult))
-    return IndexExprsLatticeStorage::top();
-  auto [joinedSourceVectorShape, joinedSourceVectorShapePriority] =
-      *joinedSVSResult;
-
-  SmallVector<NamedAttribute> priEntries;
-  priEntries.reserve(resultPriorities.size());
-  for (auto &[key, pri] : resultPriorities)
-    priEntries.emplace_back(key, IntegerAttr::get(i32, pri));
-
-  IndexExprsLatticeStorage joined(
-      DictionaryAttr::get(ctx, llvm::map_to_vector(result,
-                                                   [](auto &&pair) {
-                                                     return NamedAttribute(
-                                                         pair.first,
-                                                         pair.second);
-                                                   })),
-      DictionaryAttr::get(ctx, priEntries), joinedVectorShape,
-      joinedSourceVectorShape, joinedSourceVectorShapePriority);
-  return joined;
-}
-
-wave::IndexExprsLatticeStorage
-wave::IndexExprsLatticeStorage::meet(const IndexExprsLatticeStorage &lhs,
-                                     const IndexExprsLatticeStorage &rhs) {
-  return join(lhs, rhs);
-}
-
-void wave::IndexExprsLatticeStorage::unsafeSet(
-    const IndexExprsLatticeStorage &value) {
-  operator=(value);
-}
-
-wave::IndexExprsLatticeStorage wave::IndexExprsLatticeStorage::keepOnlySymbols(
-    llvm::ArrayRef<wave::WaveSymbolAttr> symbols) const {
-  if (isBottom() || isTop())
-    return *this;
-
-  llvm::StringSet<> symbolNames;
-  for (wave::WaveSymbolAttr symbol : symbols)
-    symbolNames.insert(symbol.getName());
-
-  llvm::SmallVector<NamedAttribute> filtered =
-      llvm::filter_to_vector(getConcreteValue(), [&](NamedAttribute attr) {
-        return symbolNames.contains(attr.getName().getValue());
-      });
-  DictionaryAttr filteredVectorShape =
-      detail::filterVectorShape(getVectorShape(), symbols);
-
-  if (filtered.empty())
-    return bottom();
-
-  MLIRContext *ctx = getConcreteValue().getContext();
-  IntegerType i32 = IntegerType::get(ctx, 32);
-  SmallVector<NamedAttribute> filteredPriorities;
-  filteredPriorities.reserve(filtered.size());
-  for (const NamedAttribute &attr : filtered)
-    filteredPriorities.emplace_back(
-        attr.getName(),
-        IntegerAttr::get(i32, getPriorityForKey(attr.getName())));
-
-  return IndexExprsLatticeStorage(DictionaryAttr::get(ctx, filtered),
-                                  DictionaryAttr::get(ctx, filteredPriorities),
-                                  filteredVectorShape, sourceVectorShape,
-                                  sourceVectorShapePriority);
-}
-
-wave::IndexExprsLatticeStorage
-wave::IndexExprsLatticeStorage::withoutIterSymbols(
-    llvm::ArrayRef<wave::WaveSymbolAttr> iterSymbols) const {
-  if (isBottom() || isTop())
-    return *this;
-
-  MLIRContext *ctx = getConcreteValue().getContext();
-  llvm::SmallVector<NamedAttribute> updated =
-      llvm::map_to_vector(getConcreteValue(), [&](NamedAttribute attr) {
-        auto value = llvm::cast<wave::WaveIndexMappingAttr>(attr.getValue());
-        for (wave::WaveSymbolAttr iterSymbol : iterSymbols) {
-          auto actualIterSymbol =
-              wave::WaveIterSymbolAttr::get(ctx, iterSymbol.getName());
-          value = value.removeInput(actualIterSymbol);
-        }
-        return NamedAttribute(attr.getName(), value);
-      });
-  IndexExprsLatticeStorage result(DictionaryAttr::get(ctx, updated),
-                                  getPriorities(), getVectorShape());
-  result.sourceVectorShape = sourceVectorShape;
-  result.sourceVectorShapePriority = sourceVectorShapePriority;
-  return result;
-}
-
-void wave::IndexExprsLatticeStorage::print(llvm::raw_ostream &os) const {
-  if (isBottom()) {
-    os << "<bottom>";
-  } else if (isTop()) {
-    os << "<top>";
-  } else {
-    os << "[pri: {";
-    bool first = true;
-    for (NamedAttribute attr : getConcreteValue()) {
-      if (!first)
-        os << ", ";
-      first = false;
-      os << attr.getName().getValue() << "="
-         << getPriorityForKey(attr.getName());
-    }
-    os << "}] " << getConcreteValue();
-    if (DictionaryAttr shape = getVectorShape())
-      os << " vectorShape: " << shape;
-    if (DictionaryAttr svs = getSourceVectorShape())
-      os << " sourceVectorShape(pri=" << getSourceVectorShapePriority()
-         << "): " << svs;
-  }
-}
-
-void wave::IndexExprsLatticeStorage::dump() const { print(llvm::errs()); }
-
-void wave::operator<<(Diagnostic &diag, const IndexExprsLatticeStorage &value) {
-  std::string str;
-  llvm::raw_string_ostream os(str);
-  value.print(os);
-  diag << os.str();
-}
-
-llvm::raw_ostream &
-llvm::operator<<(llvm::raw_ostream &os,
-                 const wave::IndexExprsLatticeStorage &lattice) {
-  lattice.print(os);
-  return os;
-}
-
-llvm::FailureOr<ChangeResult> wave::detail::identityIndexExprsPropagate(
-    llvm::ArrayRef<IndexExprsLatticeStorage> from,
-    llvm::MutableArrayRef<IndexExprsLatticeStorage> to, ValueRange toValues,
-    llvm::StringRef fromName, llvm::StringRef toName,
-    wave::EmitErrorFn emitError) {
-  ChangeResult changeResult = ChangeResult::NoChange;
-  for (auto &&[toNum, toLattice, toValue] : llvm::enumerate(to, toValues)) {
-    if (toLattice.isTop())
-      continue;
-
-    auto toTensorType = dyn_cast<WaveTensorType>(toValue.getType());
-    if (!toTensorType || !toTensorType.getFullySpecified())
-      continue;
-
-    for (auto &&[fromNum, fromLattice] : llvm::enumerate(from)) {
-      bool fromTop = false;
-
-      // If one of the from lattices reached the top, no need to keep joining,
-      // the result is known to be top. The error will have been reported
-      // already, no need to repeat it.
-      if (fromLattice.isTop()) {
-        fromTop = true;
-        toLattice = IndexExprsLatticeStorage::top();
-        changeResult = ChangeResult::Change;
-        break;
-      }
-
-      // XXX: a more efficient way would have been to join all "from" lattices
-      // first, and then join that into each "to" lattice. But this heuristic
-      // would not work in that case.
-      if (!wave::detail::shouldPropagateIndexExprs(fromLattice, toLattice,
-                                                   toValue)) {
-        LLVM_DEBUG(LDBG() << "not propagating index expressions from "
-                          << fromName << " #" << fromNum << " to " << toName
-                          << " #" << toNum << "\n";);
-        continue;
-      }
-
-      IndexExprsLatticeStorage joined =
-          IndexExprsLatticeStorage::join(toLattice, fromLattice);
-      if (joined.isTop() && !fromTop) {
-        StringRef conflictKind = "index expressions";
-        if (failed(IndexExprsLatticeStorage::getJoinedVectorShape(
-                toLattice, fromLattice)) &&
-            toLattice.getVectorShape() && fromLattice.getVectorShape())
-          conflictKind = "vector shapes";
-        else if (failed(IndexExprsLatticeStorage::getJoinedSourceVectorShape(
-                     toLattice, fromLattice)))
-          conflictKind = "source vector shapes";
-        InFlightDiagnostic diag =
-            emitError() << "conflict when propagating " << conflictKind
-                        << " from " << fromName << " #" << fromNum << " to "
-                        << toName << " #" << toNum;
-        diag.attachNote() << "original " << toName << " lattice: " << toLattice;
-        diag.attachNote() << fromName << " #" << fromNum
-                          << " lattice: " << fromLattice;
-        return diag;
-      }
-      joined = joined.keepOnlySymbols(toTensorType.getShape());
-      if (joined != toLattice) {
-        changeResult = ChangeResult::Change;
-        toLattice = joined;
-      }
-    }
-  }
-
-  return changeResult;
-}
-
-llvm::LogicalResult wave::detail::checkAndAppendIndexExpr(
-    Location loc, const IndexExprsLatticeStorage &expr,
-    const llvm::Twine &description,
-    llvm::SmallVectorImpl<Attribute> &indexExprs) {
-  if (expr.isBottom()) {
-    emitError(loc) << "failed to infer index expressions for " << description;
-    return llvm::failure();
-  }
-  if (expr.isTop()) {
-    InFlightDiagnostic diag = emitError(loc)
-                              << "conflict detected in index expressions for "
-                              << description;
-    diag.attachNote() << "PLEASE REPORT this as a bug in absence of further "
-                         "information about the conflict";
-    return llvm::failure();
-  }
-  indexExprs.push_back(expr.getConcreteValue());
-  return llvm::success();
 }
 
 // ----------------------------------------------------------------------------
