+++ water/lib/Dialect/Wave/IR/WaveOps.cpp
@@ -265,20 +265,6 @@ void wave::IterateOp::getSuccessorRegions(
   regions.emplace_back(RegionSuccessor(&getBody()));
 }
 
-llvm::FailureOr<ChangeResult> wave::IterateOp::propagateIndexExprsForward(
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage> /*operands*/,
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> /*results*/,
-    wave::EmitErrorFn /*emitError*/) {
-  llvm_unreachable("IterateOp should be handled by control flow interfaces");
-}
-
-llvm::FailureOr<ChangeResult> wave::IterateOp::propagateIndexExprsBackward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> /*operands*/,
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage> /*results*/,
-    wave::EmitErrorFn /*emitError*/) {
-  llvm_unreachable("IterateOp should be handled by control flow interfaces");
-}
-
 LogicalResult wave::IterateOp::verify() {
   if (getNumOperands() != getLoopBody()->getNumArguments()) {
     return emitOpError() << "expects the same number of operands ("
@@ -451,34 +437,6 @@ llvm::FailureOr<ChangeResult> wave::MmaOp::propagateBackward(
 
 LogicalResult wave::MmaOp::finalizeTypeInference() { return success(); }
 
-// Set the value of `lattice` to `newLattice` and return whether a change
-// happened. Note that this does NOT verify whether the lattice change goes into
-// the direction of top or bottom.
-static ChangeResult
-updateIfChanged(wave::IndexExprsLatticeStorage &lattice,
-                const wave::IndexExprsLatticeStorage &newLattice) {
-  if (newLattice == lattice)
-    return ChangeResult::NoChange;
-  lattice = newLattice;
-  return ChangeResult::Change;
-}
-
-// No propagation through MMA. The index expressions remain the same as set by
-// initialization since MMAs require very specific index expressions. If there
-// is a conflict with operands that were propagated from another MMA (other
-// operations have lower priority), it will be resolved in a separate pass after
-// the analysis completes.
-llvm::FailureOr<ChangeResult> wave::MmaOp::propagateIndexExprsForward(
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage>,
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage>, wave::EmitErrorFn) {
-  return ChangeResult::NoChange;
-}
-llvm::FailureOr<ChangeResult> wave::MmaOp::propagateIndexExprsBackward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage>,
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage>, wave::EmitErrorFn) {
-  return ChangeResult::NoChange;
-}
-
 // Check if the given type is one of the allowed types provided as template
 // arguments and report an error at the given location otherwise.
 template <typename... AllowedTypes>
@@ -566,903 +524,6 @@ static LogicalResult checkMmaTypeCompatibility(Location loc,
   }
 
   return llvm::success(success);
-}
-
-// Extract the context from the first symbol that is not null.
-static MLIRContext *getAnySymbolContext(wave::WaveSymbolAttr mSymbol,
-                                        wave::WaveSymbolAttr nSymbol,
-                                        wave::WaveSymbolAttr kSymbol) {
-  MLIRContext *context = nullptr;
-  for (wave::WaveSymbolAttr symbol : {mSymbol, nSymbol, kSymbol})
-    if (!context && symbol)
-      context = symbol.getContext();
-  assert(context && "expected at least one symbol name to be provided");
-  return context;
-}
-
-namespace {
-
-struct MmaIndexingExprBuilder;
-
-// Fluent-style API builder for index expressions of an MMA operation. See
-// MmaIndexingExprBuilder for details.
-struct MmaSingleIndexExprBuilder {
-  MmaSingleIndexExprBuilder(MmaIndexingExprBuilder &parent, bool enabled)
-      : parent(parent), enabled(enabled) {}
-
-  // Set the parameter of the index expression for the currently selected
-  // dimension.
-  MmaSingleIndexExprBuilder &offset(AffineExpr expr);
-  MmaSingleIndexExprBuilder &size(int64_t value);
-  MmaSingleIndexExprBuilder &stride(int64_t value);
-
-  // Select the dimension.
-  MmaSingleIndexExprBuilder &m();
-  MmaSingleIndexExprBuilder &n();
-  MmaSingleIndexExprBuilder &k();
-
-  // Populate the attributes with all index expressions.
-  void populate(llvm::SmallVectorImpl<NamedAttribute> &attributes) const;
-
-  MmaIndexingExprBuilder &parent;
-  AffineExpr offsetExpr, sizeExpr, strideExpr;
-  bool enabled;
-};
-
-// Fluent-style API builder for index expressions of an MMA operation. Usage:
-//   1. Create an instance of this class.
-//   2. Use `m`, `n` or `k` to select the MMA dimension to build an index
-//   expression for.
-//   3. After selecting the dimension, use `offset`, `size` or `stride` to set
-//   the corresponding quantities of the index expression.
-//   4. Proceed with the next dimension until all dimensions are set.
-//   5. Call `populate` to populate the attributes of the MMA operation.
-//
-// Example:
-//
-// ```
-//   MmaIndexingExprBuilder builder(symbols, mSymbol, nSymbol, kSymbol);
-//   builder.m().offset(offset_m).size(size_m).stride(stride_m)
-//          .n().offset(offset_n).size(size_n).stride(stride_n)
-//          .k().offset(offset_k).size(size_k).stride(stride_k)
-//          .populate(attributes);
-// ```
-struct MmaIndexingExprBuilder {
-  MmaIndexingExprBuilder(llvm::ArrayRef<Attribute> symbols,
-                         wave::WaveSymbolAttr mSymbol,
-                         wave::WaveSymbolAttr nSymbol,
-                         wave::WaveSymbolAttr kSymbol)
-      : symbols(symbols), mBuilder(*this, mSymbol != nullptr),
-        nBuilder(*this, nSymbol != nullptr),
-        kBuilder(*this, kSymbol != nullptr), mSymbol(mSymbol), nSymbol(nSymbol),
-        kSymbol(kSymbol) {
-    assert(
-        llvm::all_of(
-            symbols,
-            llvm::IsaPred<wave::WaveSymbolAttr, wave::WaveIndexSymbolAttr>) &&
-        "expected symbols to be a range of WaveSymbolAttr or "
-        "WaveIndexSymbolAttr attributes");
-  }
-
-  // Select the dimension.
-  MmaSingleIndexExprBuilder &m() { return mBuilder; }
-  MmaSingleIndexExprBuilder &n() { return nBuilder; }
-  MmaSingleIndexExprBuilder &k() { return kBuilder; }
-
-  // Populate the attributes with all index expressions.
-  void populate(llvm::SmallVectorImpl<NamedAttribute> &attributes) const {
-    MLIRContext *ctx = getAnySymbolContext(mSymbol, nSymbol, kSymbol);
-
-    auto buildMap = [&](AffineExpr expr) {
-      assert(expr &&
-             "expected offset/size/stride to be set up for all symbols");
-      return AffineMap::get(/*dimCount=*/0,
-                            /*symbolCount=*/symbols.size(), expr, ctx);
-    };
-    auto buildOne = [&](const MmaSingleIndexExprBuilder &builder) {
-      return wave::WaveIndexMappingAttr::get(
-          ctx, symbols, buildMap(builder.offsetExpr),
-          buildMap(builder.sizeExpr), buildMap(builder.strideExpr));
-    };
-
-    if (mSymbol)
-      attributes.emplace_back(mSymbol.getName(), buildOne(mBuilder));
-    if (nSymbol)
-      attributes.emplace_back(nSymbol.getName(), buildOne(nBuilder));
-    if (kSymbol)
-      attributes.emplace_back(kSymbol.getName(), buildOne(kBuilder));
-  }
-
-  llvm::ArrayRef<Attribute> symbols;
-  MmaSingleIndexExprBuilder mBuilder, nBuilder, kBuilder;
-  wave::WaveSymbolAttr mSymbol, nSymbol, kSymbol;
-};
-
-MmaSingleIndexExprBuilder &MmaSingleIndexExprBuilder::offset(AffineExpr expr) {
-  if (!enabled)
-    return *this;
-  assert(!offsetExpr && "expected offset to be set only once");
-  offsetExpr = expr;
-  return *this;
-}
-
-MmaSingleIndexExprBuilder &MmaSingleIndexExprBuilder::size(int64_t value) {
-  if (!enabled)
-    return *this;
-  assert(!sizeExpr && "expected size to be set only once");
-  sizeExpr = getAffineConstantExpr(value, offsetExpr.getContext());
-  return *this;
-}
-
-MmaSingleIndexExprBuilder &MmaSingleIndexExprBuilder::stride(int64_t value) {
-  if (!enabled)
-    return *this;
-  assert(!strideExpr && "expected stride to be set only once");
-  strideExpr = getAffineConstantExpr(value, offsetExpr.getContext());
-  return *this;
-}
-
-[[maybe_unused]] MmaSingleIndexExprBuilder &MmaSingleIndexExprBuilder::m() {
-  return parent.m();
-}
-[[maybe_unused]] MmaSingleIndexExprBuilder &MmaSingleIndexExprBuilder::n() {
-  return parent.n();
-}
-[[maybe_unused]] MmaSingleIndexExprBuilder &MmaSingleIndexExprBuilder::k() {
-  return parent.k();
-}
-void MmaSingleIndexExprBuilder::populate(
-    llvm::SmallVectorImpl<NamedAttribute> &attributes) const {
-  parent.populate(attributes);
-}
-} // namespace
-
-/// MXFP group scale factor for GFX950 and GFX1250.
-static constexpr unsigned kScaleGroupSize = 32;
-
-// Get the attribute representing the vector shape implied by the given MMA
-// operation kind for its M, N, K dimensions.
-static DictionaryAttr getMmaVectorShape(
-    Location loc, wave::WaveMmaKind kind, wave::WaveSymbolAttr mSymbol,
-    wave::WaveSymbolAttr nSymbol, wave::WaveSymbolAttr kSymbol,
-    wave::WaveSymbolAttr kScaledSymbol,
-    ArrayRef<wave::WaveSymbolAttr> batchSymbols, DictionaryAttr hwVectorShape,
-    wave::EmitDelayedErrorFn *delayedErrorEmitter) {
-  int m, n, k, kScaled;
-  switch (kind) {
-  case wave::WaveMmaKind::F32_16x16x16_F16:
-  case wave::WaveMmaKind::I32_16x16x16_I8:
-    m = 16;
-    n = 16;
-    k = 16;
-    break;
-  case wave::WaveMmaKind::F32_32x32x8_F16:
-  case wave::WaveMmaKind::I32_32x32x8_I8:
-    m = 32;
-    n = 32;
-    k = 8;
-    break;
-  case wave::WaveMmaKind::F32_16x16x32_F8:
-  case wave::WaveMmaKind::F32_16x16x32_BF16:
-  case wave::WaveMmaKind::F32_16x16x32_F16:
-  case wave::WaveMmaKind::F32_16x16x32_K8_F16:
-  case wave::WaveMmaKind::I32_16x16x32_I8:
-  case wave::WaveMmaKind::F32_16x16x32_K4_F8:
-    m = 16;
-    n = 16;
-    k = 32;
-    break;
-  case wave::WaveMmaKind::F32_32x32x16_F8:
-  case wave::WaveMmaKind::F32_32x32x16_BF16:
-  case wave::WaveMmaKind::F32_32x32x16_F16:
-  case wave::WaveMmaKind::F32_32x32x16_K8_F16:
-  case wave::WaveMmaKind::I32_32x32x16_I8:
-  case wave::WaveMmaKind::F32_32x32x16_K4_F8:
-    m = 32;
-    n = 32;
-    k = 16;
-    break;
-  case wave::WaveMmaKind::F32_16x16x128_F8F6F4:
-    m = 16;
-    n = 16;
-    k = 128;
-    kScaled = k / kScaleGroupSize;
-    break;
-  case wave::WaveMmaKind::F32_32x32x64_F8F6F4:
-    m = 32;
-    n = 32;
-    k = 64;
-    kScaled = k / kScaleGroupSize;
-    break;
-  }
-
-  MLIRContext *ctx = loc->getContext();
-  auto iAttr = [&](int64_t value) {
-    return IntegerAttr::get(IntegerType::get(ctx, 64), value);
-  };
-  SmallVector<NamedAttribute> attributes;
-  if (mSymbol)
-    attributes.emplace_back(mSymbol.getName(), iAttr(m));
-  if (nSymbol)
-    attributes.emplace_back(nSymbol.getName(), iAttr(n));
-  if (kSymbol)
-    attributes.emplace_back(kSymbol.getName(), iAttr(k));
-  if (kScaledSymbol)
-    attributes.emplace_back(kScaledSymbol.getName(), iAttr(kScaled));
-
-  if (hwVectorShape) {
-    for (NamedAttribute attr : hwVectorShape) {
-      if (mSymbol && attr.getName() == mSymbol.getName()) {
-        if (delayedErrorEmitter &&
-            cast<IntegerAttr>(attr.getValue()).getValue() != m) {
-          wave::EmitDelayedErrorFn previous = *delayedErrorEmitter;
-          *delayedErrorEmitter = [mSymbol, m,
-                                  previous](InFlightDiagnostic &diag) {
-            if (previous)
-              previous(diag);
-            diag << "overriding vector shape for " << mSymbol << " to " << m
-                 << " implied by the MMA operation";
-          };
-        }
-        continue;
-      }
-      if (nSymbol && attr.getName() == nSymbol.getName()) {
-        if (delayedErrorEmitter &&
-            cast<IntegerAttr>(attr.getValue()).getValue() != n) {
-          wave::EmitDelayedErrorFn previous = *delayedErrorEmitter;
-          *delayedErrorEmitter = [nSymbol, n,
-                                  previous](InFlightDiagnostic &diag) {
-            if (previous)
-              previous(diag);
-            diag << "overriding vector shape for " << nSymbol << " to " << n
-                 << " implied by the MMA operation";
-          };
-        }
-        continue;
-      }
-      if (kSymbol && attr.getName() == kSymbol.getName()) {
-        if (delayedErrorEmitter &&
-            cast<IntegerAttr>(attr.getValue()).getValue() != k) {
-          wave::EmitDelayedErrorFn previous = *delayedErrorEmitter;
-          *delayedErrorEmitter = [kSymbol, k,
-                                  previous](InFlightDiagnostic &diag) {
-            if (previous)
-              previous(diag);
-            diag << "overriding vector shape for " << kSymbol << " to " << k
-                 << " implied by the MMA operation";
-          };
-        }
-        continue;
-      }
-      if (kScaledSymbol && attr.getName() == kScaledSymbol.getName()) {
-        if (delayedErrorEmitter &&
-            cast<IntegerAttr>(attr.getValue()).getValue() != kScaled) {
-          wave::EmitDelayedErrorFn previous = *delayedErrorEmitter;
-          *delayedErrorEmitter = [kScaledSymbol, kScaled,
-                                  previous](InFlightDiagnostic &diag) {
-            if (previous)
-              previous(diag);
-            diag << "overriding vector shape for " << kScaledSymbol << " to "
-                 << kScaled << " implied by the Scaled MMA operation";
-          };
-        }
-        continue;
-      }
-      if (llvm::any_of(batchSymbols, [&](wave::WaveSymbolAttr symbol) {
-            return symbol.getName() == attr.getName();
-          })) {
-        attributes.push_back(attr);
-      }
-    }
-  }
-
-  return DictionaryAttr::get(ctx, attributes);
-}
-
-// Populate `attributes` with index expressions for the symbols associated with
-// M, N, K dimensions of the given MMA operation kind provided the configuration
-// of wavefronts in the workgroup. Any symbol may be omitted as long as at least
-// one is provided, e.g., for the LHS of the operation, only M and N symbols may
-// be provided. If `isAccumulator` is set, the index expressions are created for
-// the accumulator/result of an MMA, which may affect the expression for the M
-// dimension. If `isScaled` is set, the K expression uses the scale operand
-// mapping (one scale per kScaleGroupSize K-elements). If `isFP4` is set,
-// the K expression uses the FP4 data mapping. Both `isScaled` and `isFP4`
-// currently produce the same K offset formula (they are ORed together), but
-// are kept separate because they represent distinct semantic cases: `isScaled`
-// refers to E8M0 scale operands while `isFP4` refers to FP4 data operands.
-static llvm::LogicalResult populateMmaIndexingExpr(
-    wave::WaveMmaKind kind, bool isAccumulator, bool isScaled, bool isFP4,
-    llvm::ArrayRef<unsigned> wavesPerWorkgroup, int64_t threadsPerWave,
-    wave::WaveSymbolAttr mSymbol, wave::WaveSymbolAttr nSymbol,
-    wave::WaveSymbolAttr kSymbol,
-    llvm::SmallVectorImpl<NamedAttribute> &attributes) {
-  MLIRContext *ctx = getAnySymbolContext(mSymbol, nSymbol, kSymbol);
-
-  llvm::SmallVector<Attribute> symbolNames = {
-      wave::WaveIndexSymbolAttr::get(ctx, wave::WaveIndexSymbol::THREAD_0),
-      wave::WaveIndexSymbolAttr::get(ctx, wave::WaveIndexSymbol::THREAD_1),
-      wave::WaveIndexSymbolAttr::get(ctx, wave::WaveIndexSymbol::THREAD_2),
-      wave::WaveIndexSymbolAttr::get(ctx, wave::WaveIndexSymbol::GPR_NUMBER),
-  };
-  AffineExpr threadX, threadY, threadZ, gprNum;
-  bindSymbols(ctx, threadX, threadY, threadZ, gprNum);
-
-  AffineExpr linearizedThreadId =
-      threadX + threadY * wavesPerWorkgroup[0] * threadsPerWave +
-      threadZ * wavesPerWorkgroup[1] * wavesPerWorkgroup[0] * threadsPerWave;
-  AffineExpr laneId = linearizedThreadId % threadsPerWave;
-  MmaIndexingExprBuilder builder(symbolNames, mSymbol, nSymbol, kSymbol);
-
-  switch (kind) {
-  case wave::WaveMmaKind::F32_16x16x16_F16:
-  case wave::WaveMmaKind::I32_16x16x16_I8:
-    builder.m()
-        .offset(isAccumulator ? 4 * laneId.floorDiv(16) : laneId % 16)
-        .size(isAccumulator ? 4 : 1)
-        .stride(isAccumulator ? 16 : 1)
-        .n()
-        .offset(laneId % 16)
-        .size(1)
-        .stride(1)
-        .k()
-        .offset(4 * laneId.floorDiv(16))
-        .size(4)
-        .stride(1)
-        .populate(attributes);
-    return llvm::success();
-
-  case wave::WaveMmaKind::F32_32x32x8_F16:
-  case wave::WaveMmaKind::I32_32x32x8_I8:
-    builder.m()
-        .offset(isAccumulator ? (8 * gprNum.floorDiv(4) % 32) +
-                                    4 * laneId.floorDiv(32) + (gprNum % 4)
-                              : laneId % 32)
-        .size(isAccumulator ? 16 : 1)
-        .stride(isAccumulator ? 32 : 1)
-        .n()
-        .offset(laneId % 32)
-        .size(1)
-        .stride(1)
-        .k()
-        .offset(4 * laneId.floorDiv(32))
-        .size(4)
-        .stride(1)
-        .populate(attributes);
-    return llvm::success();
-
-  case wave::WaveMmaKind::F32_16x16x32_F8:
-  case wave::WaveMmaKind::F32_16x16x32_BF16:
-  case wave::WaveMmaKind::F32_16x16x32_F16:
-  case wave::WaveMmaKind::F32_16x16x32_K8_F16:
-  case wave::WaveMmaKind::I32_16x16x32_I8:
-    builder.m()
-        .offset(isAccumulator ? 4 * laneId.floorDiv(16) : laneId % 16)
-        .size(isAccumulator ? 4 : 1)
-        .stride(isAccumulator ? 16 : 1)
-        .n()
-        .offset(laneId % 16)
-        .size(1)
-        .stride(1)
-        .k()
-        .offset(8 * laneId.floorDiv(16))
-        .size(8)
-        .stride(1)
-        .populate(attributes);
-    return llvm::success();
-  case wave::WaveMmaKind::F32_16x16x32_K4_F8:
-    builder.m()
-        .offset(isAccumulator ? 4 * laneId.floorDiv(16) : laneId % 16)
-        .size(isAccumulator ? 4 : 1)
-        .stride(isAccumulator ? 16 : 1)
-        .n()
-        .offset(laneId % 16)
-        .size(1)
-        .stride(1)
-        .k()
-        .offset(16 * gprNum.floorDiv(4) + 4 * laneId.floorDiv(16) +
-                (gprNum % 4))
-        .size(8)
-        .stride(1)
-        .populate(attributes);
-    return llvm::success();
-  case wave::WaveMmaKind::F32_32x32x16_F8:
-  case wave::WaveMmaKind::F32_32x32x16_BF16:
-  case wave::WaveMmaKind::F32_32x32x16_F16:
-  case wave::WaveMmaKind::F32_32x32x16_K8_F16:
-  case wave::WaveMmaKind::I32_32x32x16_I8:
-    builder.m()
-        .offset(isAccumulator ? (8 * gprNum.floorDiv(4) % 32) +
-                                    4 * laneId.floorDiv(32) + (gprNum % 4)
-                              : laneId % 32)
-        .size(isAccumulator ? 16 : 1)
-        .stride(isAccumulator ? 32 : 1)
-        .n()
-        .offset(laneId % 32)
-        .size(1)
-        .stride(1)
-        .k()
-        .offset(8 * laneId.floorDiv(32))
-        .size(8)
-        .stride(1)
-        .populate(attributes);
-    return llvm::success();
-  case wave::WaveMmaKind::F32_32x32x16_K4_F8:
-    builder.m()
-        .offset(isAccumulator ? (8 * gprNum.floorDiv(4) % 32) +
-                                    4 * laneId.floorDiv(32) + (gprNum % 4)
-                              : laneId % 32)
-        .size(isAccumulator ? 16 : 1)
-        .stride(isAccumulator ? 32 : 1)
-        .n()
-        .offset(laneId % 32)
-        .size(1)
-        .stride(1)
-        .k()
-        .offset(8 * gprNum.floorDiv(4) + 4 * laneId.floorDiv(32) + (gprNum % 4))
-        .size(8)
-        .stride(1)
-        .populate(attributes);
-    return llvm::success();
-
-  case wave::WaveMmaKind::F32_16x16x128_F8F6F4:
-    builder.m()
-        .offset(isAccumulator ? 4 * laneId.floorDiv(16) : laneId % 16)
-        .size(isAccumulator ? 4 : 1)
-        .stride(isAccumulator ? 16 : 1)
-        .n()
-        .offset(laneId % 16)
-        .size(1)
-        .stride(1)
-        .k()
-        .offset((isScaled || isFP4)
-                    ? 32 * laneId.floorDiv(16)
-                    : 64 * gprNum.floorDiv(16) + 16 * laneId.floorDiv(16) +
-                          (gprNum % 16))
-        .size(isScaled ? 1 : 32)
-        .stride(1)
-        .populate(attributes);
-    return llvm::success();
-  case wave::WaveMmaKind::F32_32x32x64_F8F6F4:
-    // XXX: verify whether the K offset should also vary with isScaled/isFP4
-    // for this variant, similar to F32_16x16x128_F8F6F4. Currently ported
-    // directly from PyWave.
-    builder.m()
-        .offset(isAccumulator ? (8 * gprNum.floorDiv(4) % 32) +
-                                    4 * laneId.floorDiv(32) + (gprNum % 4)
-                              : laneId % 32)
-        .size(isAccumulator ? 16 : 1)
-        .stride(isAccumulator ? 32 : 1)
-        .n()
-        .offset(laneId % 32)
-        .size(1)
-        .stride(1)
-        .k()
-        .offset(32 * laneId.floorDiv(32))
-        .size(32)
-        .stride(1)
-        .populate(attributes);
-    return llvm::success();
-  }
-}
-
-/// Create per-symbol thread-independent index expressions for `indexingSymbols`
-/// given constraints on them and put them into `symbolMappings` as named pairs
-/// (symbol, index mapping attribute). Thread-independent means affected by
-/// workgroup, tiling and device constraints, and NOT affected by wave
-/// constraints or MMA shapes. The first argument indicates for which operation
-/// the constraints are being used, which is in particular necessary to only
-/// apply tiling constraints inside the relevant loops.
-template <typename RangeT>
-static void mixInThreadIndependentConstraints(
-    Operation *where, uint64_t threadsPerWave, RangeT &&indexingSymbols,
-    const llvm::DenseMap<wave::WaveSymbolAttr, llvm::SmallVector<Attribute>>
-        &symbolConstraints,
-    llvm::SmallVectorImpl<NamedAttribute> &symbolMappings) {
-
-  static_assert(
-      std::is_same_v<std::decay_t<decltype(*std::declval<RangeT>().begin())>,
-                     wave::WaveSymbolAttr>,
-      "expected a range of WaveSymbolAttr");
-
-  auto zero = AffineMap::get(/*dimCount=*/0, /*numSymbols=*/0,
-                             getAffineConstantExpr(0, where->getContext()));
-  auto one = AffineMap::get(/*dimCount=*/0, /*numSymbols=*/0,
-                            getAffineConstantExpr(1, where->getContext()));
-  for (wave::WaveSymbolAttr symbol : indexingSymbols) {
-    auto mappingIt = llvm::find_if(symbolMappings, [&](NamedAttribute attr) {
-      return attr.getName() == symbol.getName();
-    });
-    wave::WaveIndexMappingAttr mapping = [&]() {
-      if (mappingIt != symbolMappings.end())
-        return llvm::cast<wave::WaveIndexMappingAttr>(mappingIt->getValue());
-
-      // If no other mapping is present, default to (start=0, step=1, stride=1),
-      // assuming this will be replicated enough times by the expansion pass.
-      // Constraints may be applied below to amend this mapping.
-      auto mapping = wave::WaveIndexMappingAttr::get(
-          where->getContext(), /*symbols=*/{}, zero, one, one);
-      symbolMappings.emplace_back(symbol.getName(), mapping);
-      return mapping;
-    }();
-
-    auto it = symbolConstraints.find(symbol);
-    if (it == symbolConstraints.end()) {
-      continue;
-    }
-
-    // There is interaction between constraints of different kinds for the same
-    // symbol, find them all upfront.
-    wave::WorkgroupConstraintAttr workgroupConstraint;
-    wave::WaveConstraintAttr waveConstraint;
-    wave::TilingConstraintAttr tilingConstraint;
-    for (Attribute constraint : it->second) {
-      if (auto maybeWorkgroupConstraint =
-              dyn_cast<wave::WorkgroupConstraintAttr>(constraint)) {
-        workgroupConstraint = maybeWorkgroupConstraint;
-      } else if (auto maybeWaveConstraint =
-                     dyn_cast<wave::WaveConstraintAttr>(constraint)) {
-        waveConstraint = maybeWaveConstraint;
-      } else if (auto maybeTilingConstraint =
-                     dyn_cast<wave::TilingConstraintAttr>(constraint)) {
-        tilingConstraint = maybeTilingConstraint;
-      } else {
-        llvm_unreachable("unsupported constraint type");
-      }
-    }
-
-    if (tilingConstraint) {
-      // Tiling constraints should only be applied inside the corresponding
-      // parent iterate op.
-      for (Operation *parent = where->getParentOp(); parent;
-           parent = parent->getParentOp()) {
-        auto iterateOp = dyn_cast<wave::IterateOp>(parent);
-        if (!iterateOp)
-          continue;
-        wave::WaveSymbolAttr iterSymbol = iterateOp.getIterator();
-        if (iterSymbol.getName() == symbol.getName()) {
-          mapping = applyConstraint(tilingConstraint, mapping);
-          break;
-        }
-      }
-    }
-
-    if (workgroupConstraint)
-      mapping = applyConstraint(workgroupConstraint, mapping);
-
-    if (waveConstraint) {
-      assert(workgroupConstraint && "workgroup constraint must be present if a "
-                                    "wave constraint for the same symbol is");
-      mapping = applyConstraint(
-          waveConstraint, workgroupConstraint.getWorkgroupDim().getValue(),
-          threadsPerWave, mapping);
-    }
-
-    if (mappingIt != symbolMappings.end())
-      mappingIt->setValue(mapping);
-    else if (mapping)
-      symbolMappings.emplace_back(symbol.getName(), mapping);
-  }
-}
-
-// Joins the given lattice with another lattice in place and handles conflicts.
-// If the lattice is already top, or the join result does not change the
-// lattice, the function returns success. If joining produces top (i.e., an
-// undecidable or conflicting result) that differs from the original, an error
-// is emitted using the provided emitError function.
-static LogicalResult
-joinIndexExprsLatticeInPlace(wave::IndexExprsLatticeStorage &lattice,
-                             StringRef latticeName,
-                             const wave::IndexExprsLatticeStorage &other,
-                             StringRef otherName, wave::EmitErrorFn emitError) {
-  if (lattice.isTop())
-    return success();
-
-  IndexExprsLatticeStorage joined =
-      IndexExprsLatticeStorage::join(lattice, other);
-  if (joined == lattice)
-    return success();
-  // When newly reached top, report an error.
-  if (joined.isTop() && !other.isTop()) {
-    StringRef conflictKind = " index expression";
-    if (failed(IndexExprsLatticeStorage::getJoinedVectorShape(lattice, other)))
-      conflictKind = " vector shape";
-    else if (failed(IndexExprsLatticeStorage::getJoinedSourceVectorShape(
-                 lattice, other)))
-      conflictKind = " source vector shape";
-    InFlightDiagnostic diag =
-        emitError() << "conflict for " << latticeName << conflictKind
-                    << " when propagating from " << otherName << " lattice";
-    diag.attachNote() << "original " << latticeName << " lattice: " << lattice;
-    diag.attachNote() << otherName << " lattice: " << other;
-    return diag;
-  }
-
-#ifndef NDEBUG
-  assert(IndexExprsLatticeStorage::join(joined, lattice) == joined &&
-         "join should not move the lattice backward");
-  assert(IndexExprsLatticeStorage::join(joined, other) == joined &&
-         "join should not move the lattice forward");
-#endif
-
-  lattice.unsafeSet(joined);
-  return success();
-}
-
-bool wave::detail::shouldPropagateIndexExprs(
-    const wave::IndexExprsLatticeStorage &from,
-    const wave::IndexExprsLatticeStorage &to, Value toValue) {
-  DictionaryAttr sourceVectorShapes = from.getSourceVectorShape();
-  if (from.isBottom() || from.isTop() || !sourceVectorShapes || to.isTop() ||
-      to.isBottom())
-    return true;
-
-  if (toValue.getDefiningOp<wave::MmaOp>()) {
-    LLVM_DEBUG(LDBG() << "skipping update to mma");
-    return false;
-  }
-
-  auto toType = cast<WaveTensorType>(toValue.getType());
-  ArrayRef<wave::WaveSymbolAttr> toShape = toType.getShape();
-  if (from.getSourceVectorShapePriority() > 0 &&
-      !llvm::all_of(toShape, [&](wave::WaveSymbolAttr symbol) {
-        return sourceVectorShapes.get(symbol.getName()) != nullptr;
-      })) {
-    LLVM_DEBUG(LDBG() << "skipping update to " << toValue
-                      << " because source vector shape does not include all "
-                         "target dimensions");
-    return false;
-  }
-
-  SmallVector<StringAttr> nonBatchDims;
-  for (WaveSymbolAttr attr : toType.getShape()) {
-    Attribute dimSize = sourceVectorShapes.get(attr.getName());
-    if (!dimSize)
-      continue;
-    if (llvm::cast<IntegerAttr>(dimSize).getValue().getSExtValue() > 1) {
-      nonBatchDims.push_back(
-          StringAttr::get(attr.getContext(), attr.getName()));
-    }
-  }
-
-  SmallVector<StringAttr> toKeys = llvm::map_to_vector(
-      to.getConcreteValue(),
-      [](const NamedAttribute &attr) { return attr.getName(); });
-  auto isSubset = [](ArrayRef<StringAttr> subset,
-                     ArrayRef<StringAttr> superset) {
-    return llvm::all_of(subset, [&](StringAttr name) {
-      return llvm::is_contained(superset, name);
-    });
-  };
-  if ((toKeys.size() < nonBatchDims.size() &&
-       !isSubset(toKeys, nonBatchDims)) ||
-      !isSubset(nonBatchDims, toKeys)) {
-    LLVM_DEBUG(LDBG() << "skipping update to " << toValue
-                      << " because of the subset condition");
-    return false;
-  }
-
-  return true;
-}
-
-LogicalResult wave::detail::buildThreadIndependentIndexMappings(
-    Operation *op, Type type, const wave::IndexExprsAnalysisInit &initObject,
-    llvm::SmallVectorImpl<mlir::NamedAttribute> &symbolMappings) {
-  auto tensorType = dyn_cast<wave::WaveTensorType>(type);
-  if (!tensorType)
-    return failure();
-
-  ArrayRef<wave::WaveSymbolAttr> indexingSymbols = tensorType.getShape();
-  mixInThreadIndependentConstraints(
-      op, initObject.hardwareConstraint.getThreadsPerWave(), indexingSymbols,
-      initObject.symbolConstraints, symbolMappings);
-  return success();
-}
-
-// Initialize the index expression lattices for the result of the MMA operation.
-// This sets index expressions to values derived from the MMA operation kind and
-// wavefront-in-workgroup configuration (thread-dependent) as well as workgroup
-// constraints (thread-independent).
-LogicalResult MmaOp::initializeIndexExprsForward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> resultExprs,
-    const IndexExprsAnalysisInit &initObject, wave::EmitErrorFn emitError) {
-  ArrayRef<wave::WaveSymbolAttr> indexingSymbols =
-      cast<wave::WaveTensorType>(getResult().getType()).getShape();
-  SmallVector<NamedAttribute> symbolMappings;
-  symbolMappings.reserve(indexingSymbols.size());
-
-  assert(indexingSymbols.size() >= 2 &&
-         "at least 2 indexing symbols are required for MMA result");
-  wave::WaveSymbolAttr mSymbol = indexingSymbols.drop_back().back();
-  wave::WaveSymbolAttr nSymbol = indexingSymbols.back();
-
-  std::optional<wave::WaveMmaKind> mmaKind = getKind();
-  if (!mmaKind)
-    return emitError() << "MMA operation without kind attribute not supported";
-  if (llvm::failed(populateMmaIndexingExpr(
-          *mmaKind,
-          /*isAccumulator=*/true, /*isScaled=*/false, /*isFP4=*/false,
-          initObject.wavesPerBlock,
-          initObject.hardwareConstraint.getThreadsPerWave(), mSymbol, nSymbol,
-          /*kSymbol=*/nullptr, symbolMappings))) {
-    return emitError() << "MMA kind not supported by index deduction";
-  }
-
-  // Set the priority based on the order of operations: earlier MMAs have higher
-  // priority.
-  auto orderedMmas = llvm::make_filter_range(initObject.deterministicOpOrder,
-                                             llvm::IsaPred<MmaOp>);
-  int32_t priority = wave::IndexExprsLatticeStorage::kMmaPriority +
-                     std::distance(llvm::find(orderedMmas, getOperation()),
-                                   orderedMmas.end()) -
-                     1;
-  mixInThreadIndependentConstraints(
-      *this, initObject.hardwareConstraint.getThreadsPerWave(), indexingSymbols,
-      initObject.symbolConstraints, symbolMappings);
-  wave::WaveSymbolAttr kSymbol =
-      cast<wave::WaveTensorType>(getLhs().getType()).getShape().back();
-  DictionaryAttr sourceVectorShape = getMmaVectorShape(
-      getLoc(), *mmaKind, mSymbol, nSymbol, kSymbol, /*kScaledSymbol=*/nullptr,
-      indexingSymbols.drop_back(2),
-      initObject.hardwareConstraint.getVectorShapes(), nullptr);
-  return joinIndexExprsLatticeInPlace(
-      resultExprs[0], "MMA result",
-      wave::IndexExprsLatticeStorage(
-          DictionaryAttr::get(getContext(), symbolMappings), priority,
-          getMmaVectorShape(getLoc(), *mmaKind, mSymbol, nSymbol,
-                            /*kSymbol=*/nullptr, /*kScaleSymbol=*/nullptr,
-                            /*batchSymbols=*/indexingSymbols.drop_back(2),
-                            initObject.hardwareConstraint.getVectorShapes(),
-                            /*delayedErrorEmitter=*/nullptr))
-          .withSourceVectorShape(sourceVectorShape, priority),
-      "implied by MMA kind", emitError);
-}
-
-// Initialize the index expression lattices for the operands of the MMA
-// operation. This sets index expressions to values derived from the MMA
-// operation kind and wavefront-in-workgroup configuration (thread-dependent) as
-// well as workgroup constraints (thread-independent).
-LogicalResult MmaOp::initializeIndexExprsBackward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> operandExprs,
-    const wave::IndexExprsAnalysisInit &initObject, wave::EmitErrorFn emitError,
-    wave::EmitDelayedErrorFn &delayedErrorEmitter) {
-  auto resultType = llvm::cast<wave::WaveTensorType>(getResult().getType());
-  auto lhsType = llvm::cast<wave::WaveTensorType>(getLhs().getType());
-  assert(resultType.getRank() == lhsType.getRank() && lhsType.getRank() >= 2 &&
-         "at least 2D MMA operations are supported");
-  wave::WaveSymbolAttr mSymbol = resultType.getShape().drop_back().back();
-  wave::WaveSymbolAttr nSymbol = resultType.getShape().back();
-  wave::WaveSymbolAttr kSymbol = lhsType.getShape().back();
-
-  bool isFP4 = lhsType.getElementType() == Float4E2M1FNType::get(getContext());
-
-  std::optional<wave::WaveMmaKind> mmaKind = getKind();
-  if (!mmaKind)
-    return emitError() << "MMA operation without kind attribute not supported";
-
-  // Reserve space for 1 more since the list will initially contain m,n,k along
-  // with batch dimensions until we drop either m or n for each operand.
-  llvm::SmallVector<NamedAttribute> operandSymbolMappings;
-  operandSymbolMappings.reserve(lhsType.getShape().size() + 1);
-  if (llvm::failed(populateMmaIndexingExpr(
-          *mmaKind, /*isAccumulator=*/false, /*isScaled=*/false,
-          /*isFP4=*/isFP4, initObject.wavesPerBlock,
-          initObject.hardwareConstraint.getThreadsPerWave(), mSymbol, nSymbol,
-          kSymbol, operandSymbolMappings))) {
-    return emitError() << "MMA kind not supported by index deduction";
-  }
-
-  llvm::SmallVector<NamedAttribute> accumulatorSymbolMappings;
-  accumulatorSymbolMappings.reserve(resultType.getShape().size());
-  if (llvm::failed(populateMmaIndexingExpr(
-          *mmaKind,
-          /*isAccumulator=*/true, /*isScaled=*/false, /*isFP4=*/isFP4,
-          initObject.wavesPerBlock,
-          initObject.hardwareConstraint.getThreadsPerWave(), mSymbol, nSymbol,
-          nullptr, accumulatorSymbolMappings))) {
-    return emitError() << "MMA kind not supported by index deduction";
-  }
-
-  ArrayRef<wave::WaveSymbolAttr> batchSymbols =
-      resultType.getShape().drop_back(2);
-  mixInThreadIndependentConstraints(
-      *this, initObject.hardwareConstraint.getThreadsPerWave(),
-      llvm::concat<const WaveSymbolAttr>(batchSymbols,
-                                         ArrayRef{mSymbol, nSymbol, kSymbol}),
-      initObject.symbolConstraints, operandSymbolMappings);
-  mixInThreadIndependentConstraints(
-      *this, initObject.hardwareConstraint.getThreadsPerWave(),
-      llvm::concat<const WaveSymbolAttr>(batchSymbols,
-                                         ArrayRef{mSymbol, nSymbol}),
-      initObject.symbolConstraints, accumulatorSymbolMappings);
-
-  // Create the LHS and RHS mappings that are not using symbols
-  // irrelevant for them.
-  llvm::SmallVector<NamedAttribute> lhsSymbolMappings =
-      llvm::filter_to_vector(operandSymbolMappings, [&](NamedAttribute attr) {
-        return attr.getName() != nSymbol.getName();
-      });
-  llvm::SmallVector<NamedAttribute> rhsSymbolMappings =
-      llvm::filter_to_vector(operandSymbolMappings, [&](NamedAttribute attr) {
-        return attr.getName() != mSymbol.getName();
-      });
-
-  // Set the priority based on the order of operations: earlier MMAs have higher
-  // priority.
-  auto orderedMmas = llvm::make_filter_range(initObject.deterministicOpOrder,
-                                             llvm::IsaPred<MmaOp>);
-  int32_t priority = wave::IndexExprsLatticeStorage::kMmaPriority +
-                     std::distance(llvm::find(orderedMmas, getOperation()),
-                                   orderedMmas.end()) -
-                     1;
-  auto sourceVectorShape = getMmaVectorShape(
-      getLoc(), *mmaKind, mSymbol, nSymbol, kSymbol, /*kScaledSymbol=*/nullptr,
-      batchSymbols, initObject.hardwareConstraint.getVectorShapes(), nullptr);
-  if (failed(joinIndexExprsLatticeInPlace(
-          operandExprs[getLhsMutable().getOperandNumber()], "LHS",
-          wave::IndexExprsLatticeStorage(
-              DictionaryAttr::get(getContext(), lhsSymbolMappings), priority,
-              getMmaVectorShape(getLoc(), *mmaKind, mSymbol,
-                                /*nSymbol=*/nullptr, kSymbol,
-                                /*kScaleSymbol=*/nullptr, batchSymbols,
-                                initObject.hardwareConstraint.getVectorShapes(),
-                                &delayedErrorEmitter))
-              .withSourceVectorShape(sourceVectorShape, priority),
-          "implied by MMA kind", emitError)))
-    return failure();
-  if (failed(joinIndexExprsLatticeInPlace(
-          operandExprs[getRhsMutable().getOperandNumber()], "RHS",
-          wave::IndexExprsLatticeStorage(
-              DictionaryAttr::get(getContext(), rhsSymbolMappings), priority,
-              getMmaVectorShape(getLoc(), *mmaKind,
-                                /*mSymbol=*/nullptr, nSymbol, kSymbol,
-                                /*kScaleSymbol=*/nullptr, batchSymbols,
-                                initObject.hardwareConstraint.getVectorShapes(),
-                                &delayedErrorEmitter))
-              .withSourceVectorShape(sourceVectorShape, priority),
-          "implied by MMA kind", emitError)))
-    return failure();
-  if (failed(joinIndexExprsLatticeInPlace(
-          operandExprs[getAccumulatorMutable().getOperandNumber()],
-          "accumulator",
-          wave::IndexExprsLatticeStorage(
-              DictionaryAttr::get(getContext(), accumulatorSymbolMappings),
-              priority,
-              getMmaVectorShape(getLoc(), *mmaKind, mSymbol, nSymbol,
-                                /*kSymbol=*/nullptr, /*kScaleSymbol=*/nullptr,
-                                batchSymbols,
-                                initObject.hardwareConstraint.getVectorShapes(),
-                                &delayedErrorEmitter))
-              .withSourceVectorShape(sourceVectorShape, priority),
-          "implied by MMA kind", emitError)))
-    return failure();
-  return success();
-}
-
-// Special case for MMA where we also want to have index expressions
-// for the operands.
-// TODO: this shouldn't be strictly necessary in a purely MLIR flow,
-// but is kept for Python compatibility.
-std::function<void(raw_ostream &, unsigned)>
-MmaOp::getIndexExprValuesAndDescriptions(llvm::SmallVectorImpl<Value> &values) {
-  values.reserve(4);
-  llvm::append_range(values, getOperands());
-  values.push_back(getResult());
-  unsigned lhsPosition = getLhsMutable().getOperandNumber();
-  unsigned rhsPosition = getRhsMutable().getOperandNumber();
-  unsigned accumulatorPosition = getAccumulatorMutable().getOperandNumber();
-  return [lhsPosition, rhsPosition, accumulatorPosition](raw_ostream &os,
-                                                         unsigned i) {
-    assert(i < 4 && "unexpected position");
-    if (i == lhsPosition)
-      os << "lhs";
-    else if (i == rhsPosition)
-      os << "rhs";
-    else if (i == accumulatorPosition)
-      os << "accumulator";
-    else
-      os << "result";
-  };
 }
 
 LogicalResult MmaOp::verify() {
@@ -1787,266 +848,8 @@ llvm::FailureOr<ChangeResult> wave::ScaledMmaOp::propagateBackward(
 
 LogicalResult wave::ScaledMmaOp::finalizeTypeInference() { return success(); }
 
-LogicalResult wave::ScaledMmaOp::initializeIndexExprsForward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> resultExprs,
-    const IndexExprsAnalysisInit &initObject, wave::EmitErrorFn emitError) {
-  ArrayRef<wave::WaveSymbolAttr> indexingSymbols =
-      cast<wave::WaveTensorType>(getResult().getType()).getShape();
-  SmallVector<NamedAttribute> symbolMappings;
-  symbolMappings.reserve(indexingSymbols.size());
-
-  assert(indexingSymbols.size() >= 2 &&
-         "at least 2 indexing symbols are required for MMA result");
-  wave::WaveSymbolAttr mSymbol = indexingSymbols.drop_back().back();
-  wave::WaveSymbolAttr nSymbol = indexingSymbols.back();
-
-  std::optional<wave::WaveMmaKind> mmaKind = getKind();
-  if (!mmaKind)
-    return emitError() << "scaled MMA without kind attribute not supported";
-  // The result (accumulator) is f32 and is not scaled, so isScaled=false.
-  if (llvm::failed(populateMmaIndexingExpr(
-          *mmaKind, /*isAccumulator=*/true, /*isScaled=*/false, /*isFP4=*/false,
-          initObject.wavesPerBlock,
-          initObject.hardwareConstraint.getThreadsPerWave(), mSymbol, nSymbol,
-          /*kSymbol=*/nullptr, symbolMappings))) {
-    return emitError() << "scaled MMA kind not supported by index deduction";
-  }
-
-  // Set the priority based on the order of operations: earlier scaled MMAs have
-  // higher priority.
-  auto orderedScaledMmas = llvm::make_filter_range(
-      initObject.deterministicOpOrder, llvm::IsaPred<ScaledMmaOp>);
-  int32_t priority =
-      wave::IndexExprsLatticeStorage::kMmaPriority +
-      std::distance(llvm::find(orderedScaledMmas, getOperation()),
-                    orderedScaledMmas.end()) -
-      1;
-  mixInThreadIndependentConstraints(
-      *this, initObject.hardwareConstraint.getThreadsPerWave(), indexingSymbols,
-      initObject.symbolConstraints, symbolMappings);
-  return joinIndexExprsLatticeInPlace(
-      resultExprs[0], "scaled MMA result",
-      wave::IndexExprsLatticeStorage(
-          DictionaryAttr::get(getContext(), symbolMappings), priority,
-          getMmaVectorShape(getLoc(), *mmaKind, mSymbol, nSymbol,
-                            /*kSymbol=*/nullptr, /*kScaleSymbol=*/nullptr,
-                            /*batchSymbols=*/indexingSymbols.drop_back(2),
-                            initObject.hardwareConstraint.getVectorShapes(),
-                            /*delayedErrorEmitter=*/nullptr)),
-      "implied by scaled MMA kind", emitError);
-}
-
-llvm::FailureOr<ChangeResult> wave::ScaledMmaOp::propagateIndexExprsForward(
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage>,
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage>, wave::EmitErrorFn) {
-  return ChangeResult::NoChange;
-}
-
-llvm::FailureOr<ChangeResult> wave::ScaledMmaOp::propagateIndexExprsBackward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage>,
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage>, wave::EmitErrorFn) {
-  return ChangeResult::NoChange;
-}
-
-LogicalResult wave::ScaledMmaOp::initializeIndexExprsBackward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> operandExprs,
-    const wave::IndexExprsAnalysisInit &initObject, wave::EmitErrorFn emitError,
-    wave::EmitDelayedErrorFn &delayedErrorEmitter) {
-  auto resultType = llvm::cast<wave::WaveTensorType>(getResult().getType());
-  auto lhsType = llvm::cast<wave::WaveTensorType>(getLhs().getType());
-  assert(resultType.getRank() == lhsType.getRank() && lhsType.getRank() >= 2 &&
-         "at least 2D scaled MMA operations are supported");
-  wave::WaveSymbolAttr mSymbol = resultType.getShape().drop_back().back();
-  wave::WaveSymbolAttr nSymbol = resultType.getShape().back();
-  wave::WaveSymbolAttr kSymbol = lhsType.getShape().back();
-
-  auto lhsScaleType = llvm::cast<wave::WaveTensorType>(getLhsScale().getType());
-  wave::WaveSymbolAttr kScaleSymbol = lhsScaleType.getShape().back();
-
-  std::optional<wave::WaveMmaKind> mmaKind = getKind();
-  if (!mmaKind)
-    return emitError() << "scaled MMA without kind attribute not supported";
-
-  // Data operand index expressions (lhs, rhs).
-  llvm::SmallVector<NamedAttribute> operandSymbolMappings;
-  operandSymbolMappings.reserve(lhsType.getShape().size() + 1);
-  // The result (accumulator) is f32 and is not scaled, so isScaled=false.
-  if (llvm::failed(populateMmaIndexingExpr(
-          *mmaKind, /*isAccumulator=*/false, /*isScaled=*/false,
-          /*isFP4=*/false, initObject.wavesPerBlock,
-          initObject.hardwareConstraint.getThreadsPerWave(), mSymbol, nSymbol,
-          kSymbol, operandSymbolMappings))) {
-    return emitError() << "scaled MMA kind not supported by index deduction";
-  }
-
-  // Scale operand index expressions (lhs_scale, rhs_scale).
-  llvm::SmallVector<NamedAttribute> scaleSymbolMappings;
-  scaleSymbolMappings.reserve(lhsScaleType.getShape().size() + 1);
-  if (llvm::failed(populateMmaIndexingExpr(
-          *mmaKind, /*isAccumulator=*/false, /*isScaled=*/true,
-          /*isFP4=*/false, initObject.wavesPerBlock,
-          initObject.hardwareConstraint.getThreadsPerWave(), mSymbol, nSymbol,
-          kScaleSymbol, scaleSymbolMappings))) {
-    return emitError()
-           << "scaled MMA kind not supported by scale index deduction";
-  }
-
-  // Accumulator index expressions.
-  llvm::SmallVector<NamedAttribute> accumulatorSymbolMappings;
-  accumulatorSymbolMappings.reserve(resultType.getShape().size());
-  // Accumulator type is f32 and does not get scaled
-  if (llvm::failed(populateMmaIndexingExpr(
-          *mmaKind, /*isAccumulator=*/true, /*isScaled=*/false,
-          /*isFP4=*/false, initObject.wavesPerBlock,
-          initObject.hardwareConstraint.getThreadsPerWave(), mSymbol, nSymbol,
-          nullptr, accumulatorSymbolMappings))) {
-    return emitError() << "scaled MMA kind not supported by index deduction";
-  }
-
-  ArrayRef<wave::WaveSymbolAttr> batchSymbols =
-      resultType.getShape().drop_back(2);
-
-  // Mix in thread-independent constraints for data operands.
-  mixInThreadIndependentConstraints(
-      *this, initObject.hardwareConstraint.getThreadsPerWave(),
-      llvm::concat<const WaveSymbolAttr>(batchSymbols,
-                                         ArrayRef{mSymbol, nSymbol, kSymbol}),
-      initObject.symbolConstraints, operandSymbolMappings);
-
-  // Mix in thread-independent constraints for scale operands.
-  mixInThreadIndependentConstraints(
-      *this, initObject.hardwareConstraint.getThreadsPerWave(),
-      llvm::concat<const WaveSymbolAttr>(
-          batchSymbols, ArrayRef{mSymbol, nSymbol, kScaleSymbol}),
-      initObject.symbolConstraints, scaleSymbolMappings);
-
-  // Mix in thread-independent constraints for accumulator.
-  mixInThreadIndependentConstraints(
-      *this, initObject.hardwareConstraint.getThreadsPerWave(),
-      llvm::concat<const WaveSymbolAttr>(batchSymbols,
-                                         ArrayRef{mSymbol, nSymbol}),
-      initObject.symbolConstraints, accumulatorSymbolMappings);
-
-  // LHS: filter out N symbol.
-  llvm::SmallVector<NamedAttribute> lhsSymbolMappings =
-      llvm::filter_to_vector(operandSymbolMappings, [&](NamedAttribute attr) {
-        return attr.getName() != nSymbol.getName();
-      });
-  // RHS: filter out M symbol.
-  llvm::SmallVector<NamedAttribute> rhsSymbolMappings =
-      llvm::filter_to_vector(operandSymbolMappings, [&](NamedAttribute attr) {
-        return attr.getName() != mSymbol.getName();
-      });
-  // LHS scale: filter out N symbol.
-  llvm::SmallVector<NamedAttribute> lhsScaleSymbolMappings =
-      llvm::filter_to_vector(scaleSymbolMappings, [&](NamedAttribute attr) {
-        return attr.getName() != nSymbol.getName();
-      });
-  // RHS scale: filter out M symbol.
-  llvm::SmallVector<NamedAttribute> rhsScaleSymbolMappings =
-      llvm::filter_to_vector(scaleSymbolMappings, [&](NamedAttribute attr) {
-        return attr.getName() != mSymbol.getName();
-      });
-
-  // Set the priority based on the order of operations: earlier scaled MMAs have
-  // higher priority.
-  auto orderedScaledMmas = llvm::make_filter_range(
-      initObject.deterministicOpOrder, llvm::IsaPred<ScaledMmaOp>);
-  int32_t priority =
-      wave::IndexExprsLatticeStorage::kMmaPriority +
-      std::distance(llvm::find(orderedScaledMmas, getOperation()),
-                    orderedScaledMmas.end()) -
-      1;
-  if (failed(joinIndexExprsLatticeInPlace(
-          operandExprs[getLhsMutable().getOperandNumber()], "LHS",
-          wave::IndexExprsLatticeStorage(
-              DictionaryAttr::get(getContext(), lhsSymbolMappings), priority,
-              getMmaVectorShape(getLoc(), *mmaKind, mSymbol,
-                                /*nSymbol=*/nullptr, kSymbol,
-                                /*kScaleSymbol=*/nullptr, batchSymbols,
-                                initObject.hardwareConstraint.getVectorShapes(),
-                                &delayedErrorEmitter)),
-          "implied by scaled MMA kind", emitError)))
-    return failure();
-  if (failed(joinIndexExprsLatticeInPlace(
-          operandExprs[getLhsScaleMutable().getOperandNumber()], "LHS scale",
-          wave::IndexExprsLatticeStorage(
-              DictionaryAttr::get(getContext(), lhsScaleSymbolMappings),
-              priority,
-              getMmaVectorShape(getLoc(), *mmaKind, mSymbol,
-                                /*nSymbol=*/nullptr, /*kSymbol=*/nullptr,
-                                /*kScaleSymbol=*/kScaleSymbol, batchSymbols,
-                                initObject.hardwareConstraint.getVectorShapes(),
-                                &delayedErrorEmitter)),
-          "implied by scaled MMA kind", emitError)))
-    return failure();
-  if (failed(joinIndexExprsLatticeInPlace(
-          operandExprs[getRhsMutable().getOperandNumber()], "RHS",
-          wave::IndexExprsLatticeStorage(
-              DictionaryAttr::get(getContext(), rhsSymbolMappings), priority,
-              getMmaVectorShape(getLoc(), *mmaKind,
-                                /*mSymbol=*/nullptr, nSymbol, kSymbol,
-                                /*kScaleSymbol=*/nullptr, batchSymbols,
-                                initObject.hardwareConstraint.getVectorShapes(),
-                                &delayedErrorEmitter)),
-          "implied by scaled MMA kind", emitError)))
-    return failure();
-  if (failed(joinIndexExprsLatticeInPlace(
-          operandExprs[getRhsScaleMutable().getOperandNumber()], "RHS scale",
-          wave::IndexExprsLatticeStorage(
-              DictionaryAttr::get(getContext(), rhsScaleSymbolMappings),
-              priority,
-              getMmaVectorShape(getLoc(), *mmaKind,
-                                /*mSymbol=*/nullptr, nSymbol,
-                                /*kSymbol=*/nullptr, kScaleSymbol, batchSymbols,
-                                initObject.hardwareConstraint.getVectorShapes(),
-                                &delayedErrorEmitter)),
-          "implied by scaled MMA kind", emitError)))
-    return failure();
-  if (failed(joinIndexExprsLatticeInPlace(
-          operandExprs[getAccumulatorMutable().getOperandNumber()],
-          "accumulator",
-          wave::IndexExprsLatticeStorage(
-              DictionaryAttr::get(getContext(), accumulatorSymbolMappings),
-              priority,
-              getMmaVectorShape(getLoc(), *mmaKind, mSymbol, nSymbol,
-                                /*kSymbol=*/nullptr, /*kScaleSymbol=*/nullptr,
-                                batchSymbols,
-                                initObject.hardwareConstraint.getVectorShapes(),
-                                &delayedErrorEmitter)),
-          "implied by scaled MMA kind", emitError)))
-    return failure();
-  return llvm::success();
-}
-
-std::function<void(raw_ostream &, unsigned)>
-wave::ScaledMmaOp::getIndexExprValuesAndDescriptions(
-    llvm::SmallVectorImpl<Value> &values) {
-  values.reserve(6);
-  llvm::append_range(values, getOperands());
-  values.push_back(getResult());
-  unsigned lhsPos = getLhsMutable().getOperandNumber();
-  unsigned lhsScalePos = getLhsScaleMutable().getOperandNumber();
-  unsigned rhsPos = getRhsMutable().getOperandNumber();
-  unsigned rhsScalePos = getRhsScaleMutable().getOperandNumber();
-  unsigned accPos = getAccumulatorMutable().getOperandNumber();
-  return [lhsPos, lhsScalePos, rhsPos, rhsScalePos, accPos](raw_ostream &os,
-                                                            unsigned i) {
-    assert(i < 6 && "unexpected position");
-    if (i == lhsPos)
-      os << "lhs";
-    else if (i == lhsScalePos)
-      os << "lhs scale";
-    else if (i == rhsPos)
-      os << "rhs";
-    else if (i == rhsScalePos)
-      os << "rhs scale";
-    else if (i == accPos)
-      os << "accumulator";
-    else
-      os << "result";
-  };
-}
+/// Number of K elements that share one scale factor for F8F6F4 scaled MMA.
+static constexpr unsigned kScaledMmaKScaleGroupSize = 32;
 
 /// Compute the expected elements per thread for a specific scaled MMA operand.
 /// Handles LHS, LHS scale, RHS, RHS scale, and accumulator operands.
@@ -2062,11 +865,11 @@ static llvm::FailureOr<unsigned> computeScaledMmaElementsPerThread(
   if (operandIndex == op.getLhsMutable().getOperandNumber()) {
     totalElements = spec.m * spec.k;
   } else if (operandIndex == op.getLhsScaleMutable().getOperandNumber()) {
-    totalElements = spec.m * (spec.k / kScaleGroupSize);
+    totalElements = spec.m * (spec.k / kScaledMmaKScaleGroupSize);
   } else if (operandIndex == op.getRhsMutable().getOperandNumber()) {
     totalElements = spec.n * spec.k;
   } else if (operandIndex == op.getRhsScaleMutable().getOperandNumber()) {
-    totalElements = spec.n * (spec.k / kScaleGroupSize);
+    totalElements = spec.n * (spec.k / kScaledMmaKScaleGroupSize);
   } else if (operandIndex == op.getAccumulatorMutable().getOperandNumber()) {
     totalElements = spec.m * spec.n;
   } else {
@@ -2877,286 +1680,6 @@ llvm::FailureOr<ChangeResult> wave::WriteOp::propagateElementsPerThreadBackward(
       valueOnly, "elements_per_thread attribute", "", "operand", errs);
 }
 
-// Propagate index expressions forward from the operands to the result of the
-// WriteOp. Since WriteOp has no results, this is a no-op.
-llvm::FailureOr<ChangeResult> wave::WriteOp::propagateIndexExprsForward(
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage> operandExprs,
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> resultExprs,
-    wave::EmitErrorFn emitError) {
-  // WriteOp has no results, so just return NoChange.
-  return ChangeResult::NoChange;
-}
-
-/// Computes the vector stride for each dimension: stride[i] is the product of
-/// vector shapes for dimensions i+1 .. rank-1 (so the last dimension has
-/// stride 1). A dimension is contiguous iff its stride is 1.
-static FailureOr<SmallVector<int64_t>>
-getVectorStrides(wave::WaveTensorType tensorType,
-                 HardwareConstraintAttr hardwareConstraint) {
-  assert(tensorType.getFullySpecified() &&
-         "expected fully-specified tensor type");
-  DictionaryAttr vectorShapes = hardwareConstraint.getVectorShapes();
-  if (!vectorShapes)
-    return failure();
-  int64_t rank = tensorType.getRank();
-  SmallVector<int64_t> strides(rank);
-  if (rank == 0)
-    return strides;
-  strides[rank - 1] = 1;
-  for (int64_t i = rank - 2; i >= 0; --i) {
-    Attribute vectorShape =
-        vectorShapes.get(tensorType.getShape()[i + 1].getName());
-    if (!vectorShape)
-      return failure();
-    int64_t shape = cast<IntegerAttr>(vectorShape).getValue().getSExtValue();
-    strides[i] = shape * strides[i + 1];
-  }
-  return strides;
-}
-
-LogicalResult WriteOp::initializeIndexExprsBackward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> operandExprs,
-    const wave::IndexExprsAnalysisInit &initObject, wave::EmitErrorFn emitError,
-    wave::EmitDelayedErrorFn &delayedErrorEmitter) {
-
-  // TODO: figure out how to propagate elements per threads from constraints to
-  // operations while avoiding the clash with index sequences. When propagating,
-  // we don't have sequences yet.
-  WaveTensorType tensorType = cast<WaveTensorType>(getValueToStore().getType());
-  HardwareConstraintAttr hardwareConstraint = initObject.hardwareConstraint;
-
-  assert(tensorType.getFullySpecified());
-  FailureOr<SmallVector<int64_t>> stridesOr =
-      getVectorStrides(tensorType, hardwareConstraint);
-  // XXX: don't report this error immediately since we may be able to proceed
-  // without it, e.g., when index expressions may be propagated from
-  // operations with higher priority operations to this one. This is a
-  // questionable design choice carried over from the initial Python
-  // prototype, but is needed for initial consistency. Consider revising.
-  if (failed(stridesOr)) {
-    delayedErrorEmitter = [](InFlightDiagnostic &diag) {
-      diag << "couldn't find vector shapes in the contiguity check";
-    };
-    return success();
-  }
-  llvm::ArrayRef<int64_t> strides = *stridesOr;
-  // XXX: pywave confusingly calls this "contiguous" but it is actually the
-  // dimension along which SIMD vectorization is applied, i.e., the deepest
-  // dimension for which the per-thread vector shape is not 1 or, alternatively,
-  // the product of vector shapes for trailing dimensions remains 1.
-  int64_t vectorizedDimPos = -1;
-  for (int64_t i = 0, e = tensorType.getRank(); i < e; ++i)
-    if (strides[i] == 1) {
-      vectorizedDimPos = i;
-      break;
-    }
-
-  SmallVector<NamedAttribute> indexMappings;
-  for (int64_t i = 0, e = tensorType.getRank(); i < e; ++i) {
-    AffineExpr elementsPerThread = nullptr;
-    bool isVectorized = (i == vectorizedDimPos);
-
-    // The absence of constraints for a dimension means it is not mapped to
-    // workgroups/wave/items, so there is nothing to do here.
-    // We expect it to be handled by thread-independent constraints setting
-    // the default (0, 1, 1) index expression or following the tiling
-    // constraint.
-    auto it = initObject.symbolConstraints.find(tensorType.getShape()[i]);
-    if (it == initObject.symbolConstraints.end())
-      continue;
-    auto wgConstraintIt =
-        llvm::find_if(it->second, llvm::IsaPred<WorkgroupConstraintAttr>);
-    if (wgConstraintIt == it->second.end())
-      continue;
-    WorkgroupConstraintAttr wgConstraint =
-        cast<WorkgroupConstraintAttr>(*wgConstraintIt);
-
-    // The innermost dimension with vectorized size other than 1 is the one we
-    // want to vectorize along.
-    std::optional<int64_t> opElementsPerThread = getElementsPerThread();
-    SmallVector<Attribute> symbols;
-    if (isVectorized) {
-      if (opElementsPerThread) {
-        elementsPerThread =
-            getAffineConstantExpr(*opElementsPerThread, getContext());
-      } else {
-        AffineMap tileSizeMap = wgConstraint.getTileSize().getMap();
-        assert(tileSizeMap.getNumResults() == 1 &&
-               "expected a single expression in tile size affine map");
-        unsigned numThreads = [&]() {
-          switch (wgConstraint.getWorkgroupDim().getValue()) {
-          case WaveWorkgroupDim::X:
-            return initObject.wavesPerBlock[0] *
-                   initObject.hardwareConstraint.getThreadsPerWave();
-          case WaveWorkgroupDim::Y:
-            return initObject.wavesPerBlock[1];
-          case WaveWorkgroupDim::Z:
-            return initObject.wavesPerBlock[2];
-          }
-        }();
-
-        elementsPerThread = tileSizeMap.getResult(0).ceilDiv(numThreads);
-        llvm::append_range(symbols, wgConstraint.getTileSize().getSymbols());
-      }
-    } else {
-      elementsPerThread = getAffineConstantExpr(1, getContext());
-    }
-
-    int64_t stride = strides[i];
-    assert(stride > 0 && "stride should be positive");
-
-    WaveIndexSymbol threadSymbol = [&]() {
-      switch (wgConstraint.getWorkgroupDim().getValue()) {
-      case WaveWorkgroupDim::X:
-        return WaveIndexSymbol::THREAD_0;
-      case WaveWorkgroupDim::Y:
-        return WaveIndexSymbol::THREAD_1;
-      case WaveWorkgroupDim::Z:
-        return WaveIndexSymbol::THREAD_2;
-      }
-    }();
-    WaveIndexSymbolAttr threadSymbolAttr =
-        WaveIndexSymbolAttr::get(getContext(), threadSymbol);
-    symbols.push_back(threadSymbolAttr);
-
-    AffineExpr startExpr =
-        getAffineSymbolExpr(symbols.size() - 1, getContext());
-    if (wgConstraint.getWorkgroupDim().getValue() == WaveWorkgroupDim::X) {
-      startExpr = startExpr % hardwareConstraint.getThreadsPerWave();
-    } else {
-      // TODO: in pywave, we always do `startExpr % threadsPerWave` where
-      // threadsPerWave == 1 for workgroup dims other than X, making it
-      // always zero. It mentions an assumption about the (64, 1, 1) thread
-      // shape, but it is unclear whether that assumption always holds.
-      // It looks like the intention for this was to express lane ID rather
-      // than thread ID, but it is unclear how it accounts for multiple
-      // wavefronts running in parallel.
-      startExpr = getAffineConstantExpr(0, getContext());
-    }
-
-    auto indexMapping = WaveIndexMappingAttr::get(
-        getContext(), symbols,
-        AffineMap::get(/*dimCount=*/0, symbols.size(),
-                       startExpr * elementsPerThread),
-        AffineMap::get(/*dimCount=*/0, symbols.size(), elementsPerThread),
-        AffineMap::get(/*dimCount=*/0, symbols.size(),
-                       getAffineConstantExpr(stride, getContext())));
-    indexMappings.emplace_back(tensorType.getShape()[i].getName(),
-                               indexMapping);
-  }
-  mixInThreadIndependentConstraints(
-      *this, initObject.hardwareConstraint.getThreadsPerWave(),
-      tensorType.getShape(), initObject.symbolConstraints, indexMappings);
-  if (failed(joinIndexExprsLatticeInPlace(
-          operandExprs[getValueToStoreMutable().getOperandNumber()],
-          "value to store",
-          IndexExprsLatticeStorage(
-              DictionaryAttr::get(getContext(), indexMappings),
-              IndexExprsLatticeStorage::kWritePriority,
-              detail::filterVectorShape(
-                  hardwareConstraint.getVectorShapes(),
-                  cast<WaveTensorType>(getValueToStore().getType())
-                      .getShape())),
-          "implied by write operation", emitError)))
-    return failure();
-  if (failed(joinIndexExprsLatticeInPlace(
-          operandExprs[getMemoryMutable().getOperandNumber()], "memory",
-          IndexExprsLatticeStorage(
-              DictionaryAttr::get(getContext(), indexMappings),
-              IndexExprsLatticeStorage::kWritePriority,
-              detail::filterVectorShape(
-                  hardwareConstraint.getVectorShapes(),
-                  cast<WaveTensorType>(getMemory().getType()).getShape())),
-          "implied by write operation", emitError)))
-    return failure();
-
-  return success();
-}
-
-// Propagating "sideways" between operands, but only if this would not result
-// in conflicts.
-llvm::FailureOr<ChangeResult> wave::WriteOp::propagateIndexExprsBackward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> operandExprs,
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage> resultExprs,
-    wave::EmitErrorFn emitError) {
-  // XXX: If both are propagating from MMAs, temporarily assume equal priority
-  // (ignore order of MMAs). If sideways propagation would result in a new
-  // conflict in this case, don't propagate. This is a questionable design
-  // carried over from the initial Python prototype.
-  auto forceMmaPriority = [](const IndexExprsLatticeStorage &lattice) {
-    if (lattice.isBottom() || lattice.isTop())
-      return lattice;
-    DictionaryAttr priorityDict = lattice.getPriorities();
-    bool needsCapping =
-        priorityDict && llvm::any_of(priorityDict, [](NamedAttribute na) {
-          return llvm::cast<IntegerAttr>(na.getValue()).getInt() >
-                 IndexExprsLatticeStorage::kMmaPriority;
-        });
-    if (!needsCapping)
-      return lattice;
-    MLIRContext *ctx = priorityDict.getContext();
-    IntegerType i32 = IntegerType::get(ctx, 32);
-    SmallVector<NamedAttribute> capped;
-    capped.reserve(priorityDict.size());
-    for (NamedAttribute na : priorityDict) {
-      int32_t priority = llvm::cast<IntegerAttr>(na.getValue()).getInt();
-      capped.emplace_back(
-          na.getName(),
-          IntegerAttr::get(
-              i32, std::min(priority, IndexExprsLatticeStorage::kMmaPriority)));
-    }
-    IndexExprsLatticeStorage result(lattice.getConcreteValue(),
-                                    DictionaryAttr::get(ctx, capped),
-                                    lattice.getVectorShape());
-    return result.withSourceVectorShape(
-        lattice.getSourceVectorShape(),
-        std::min(lattice.getSourceVectorShapePriority(),
-                 IndexExprsLatticeStorage::kMmaPriority));
-  };
-  IndexExprsLatticeStorage lhs = forceMmaPriority(operandExprs[0]);
-  IndexExprsLatticeStorage rhs = forceMmaPriority(operandExprs[1]);
-  auto joined = IndexExprsLatticeStorage::join(lhs, rhs);
-  if (joined.isTop() && !(lhs.isTop() || rhs.isTop())) {
-    return ChangeResult::NoChange;
-  }
-
-  // Re-join with the original expressions to get the right priority.
-  joined = IndexExprsLatticeStorage::join(operandExprs[0], operandExprs[1]);
-
-  unsigned valueToStoreOperandNumber =
-      getValueToStoreMutable().getOperandNumber();
-  unsigned memoryOperandNumber = getMemoryMutable().getOperandNumber();
-  ChangeResult changeResult = ChangeResult::NoChange;
-  if (operandExprs[valueToStoreOperandNumber] != joined &&
-      wave::detail::shouldPropagateIndexExprs(
-          operandExprs[valueToStoreOperandNumber], joined, getValueToStore())) {
-    operandExprs[valueToStoreOperandNumber] = joined;
-    changeResult = ChangeResult::Change;
-  }
-  if (operandExprs[memoryOperandNumber] != joined &&
-      wave::detail::shouldPropagateIndexExprs(operandExprs[memoryOperandNumber],
-                                              joined, getMemory())) {
-    operandExprs[memoryOperandNumber] = joined;
-    changeResult = ChangeResult::Change;
-  }
-  return changeResult;
-}
-
-// Special case for WriteOp where we want an index expression even
-// though it doesn't have results.
-// TODO: this shouldn't be necessary in a purely MLIR form since
-// mappings are a property of the SSA value (conversely, changing the
-// mapping should create a new value), but keeping for compatibility.
-std::function<void(raw_ostream &, unsigned)>
-wave::WriteOp::getIndexExprValuesAndDescriptions(
-    llvm::SmallVectorImpl<Value> &values) {
-  values.push_back(getValueToStore());
-  return [](raw_ostream &os, unsigned i) {
-    assert(i == 0 && "unexpected position");
-    os << "value to store";
-  };
-}
-
 //-----------------------------------------------------------------------------
 // YieldOp
 //-----------------------------------------------------------------------------
@@ -3352,158 +1875,6 @@ LogicalResult wave::BitcastOp::verify() {
   }
 
   return success();
-}
-
-// Remap the index expression lattice for bitcast: non-scaled dimensions pass
-// through as identity, and the scaled dimension gets its symbol renamed and its
-// step (and start/stride if present) scaled by the element bitwidth ratio.
-// \p scaledDim is the index of the dimension that carries the scaling.
-static IndexExprsLatticeStorage
-scaleBitcastIndexExprs(const IndexExprsLatticeStorage &inputLattice,
-                       ArrayRef<wave::WaveSymbolAttr> fromShape,
-                       ArrayRef<wave::WaveSymbolAttr> toShape,
-                       unsigned fromBits, unsigned toBits, unsigned scaledDim,
-                       MLIRContext *ctx) {
-  if (inputLattice.isBottom() || inputLattice.isTop())
-    return inputLattice;
-
-  assert(fromShape.size() == toShape.size() &&
-         "bitcast shapes must have equal rank");
-
-  DictionaryAttr inputDict = inputLattice.getConcreteValue();
-
-  unsigned ratio =
-      (fromBits > toBits) ? (fromBits / toBits) : (toBits / fromBits);
-  bool scaleUp = fromBits > toBits;
-
-  SmallVector<NamedAttribute> newMappings;
-  newMappings.reserve(inputDict.size());
-
-  WaveSymbolAttr fromScaledSym = fromShape[scaledDim];
-  WaveSymbolAttr toScaledSym = toShape[scaledDim];
-
-  for (NamedAttribute namedAttr : inputDict) {
-    StringRef symName = namedAttr.getName().getValue();
-    auto mapping = llvm::cast<WaveIndexMappingAttr>(namedAttr.getValue());
-
-    // Non-scaled dimensions: pass through unchanged.
-    if (symName != fromScaledSym.getName()) {
-      newMappings.push_back(namedAttr);
-      continue;
-    }
-
-    // Scaled dimension: rename the symbol key and scale the step by the
-    // bitwidth ratio.  If the step has no thread dimensions (thread-independent
-    // initialization), skip this lattice to avoid producing a zero-step mapping
-    // that would conflict with a real propagated value.
-    AffineMap stepMap = mapping.getStep();
-    if (!stepMap || (stepMap.getNumDims() == 0 && stepMap.getNumSymbols() == 0))
-      return IndexExprsLatticeStorage::bottom();
-
-    auto scaleMap = [&](AffineMap map) -> AffineMap {
-      AffineExpr scaled =
-          scaleUp ? map.getResult(0) * ratio : map.getResult(0).floorDiv(ratio);
-      return AffineMap::get(map.getNumDims(), map.getNumSymbols(), scaled, ctx);
-    };
-
-    AffineMap newStep = scaleMap(stepMap);
-    auto newMapping =
-        WaveIndexMappingAttr::get(ctx, mapping.getSymbols(), mapping.getStart(),
-                                  newStep, mapping.getStride());
-
-    newMappings.push_back(NamedAttribute(
-        StringAttr::get(ctx, toScaledSym.getName()), newMapping));
-  }
-
-  if (newMappings.empty())
-    return IndexExprsLatticeStorage::bottom();
-
-  return IndexExprsLatticeStorage(DictionaryAttr::get(ctx, newMappings),
-                                  inputLattice.getPriorities(),
-                                  inputLattice.getVectorShape());
-}
-
-// Shared propagation logic for BitcastOp index expressions in both directions.
-// \p fromExprs is the source lattice, \p toExprs is the destination to update.
-// \p fromType / \p toType are the wave tensor types on the respective sides.
-// \p fromBits / \p toBits are the element bitwidths on the respective sides.
-static llvm::FailureOr<ChangeResult> propagateBitcastIndexExprs(
-    wave::BitcastOp op, llvm::ArrayRef<IndexExprsLatticeStorage> fromExprs,
-    llvm::MutableArrayRef<IndexExprsLatticeStorage> toExprs,
-    WaveTensorType fromType, WaveTensorType toType, unsigned fromBits,
-    unsigned toBits, StringRef fromName, StringRef toName, mlir::Value toValue,
-    wave::EmitErrorFn emitError) {
-  if (!fromType || !fromType.getFullySpecified() || !toType ||
-      !toType.getFullySpecified())
-    return ChangeResult::NoChange;
-
-  if (fromBits == toBits)
-    return wave::detail::identityIndexExprsPropagate(
-        fromExprs, toExprs, toValue, fromName, toName, emitError);
-
-  // Detect which dimension is scaled using the hyperparameters.
-  SmallVector<unsigned> scaledDims = getScaledDimensions(op);
-  std::optional<unsigned> scaledDim =
-      scaledDims.empty() ? std::nullopt : std::optional(scaledDims[0]);
-
-  if (!scaledDim) {
-    InFlightDiagnostic diag = emitError()
-                              << "could not determine scaled dimension for "
-                                 "BitcastOp with differing bitwidths";
-    return diag;
-  }
-
-  IndexExprsLatticeStorage scaled = scaleBitcastIndexExprs(
-      fromExprs[0], fromType.getShape(), toType.getShape(), fromBits, toBits,
-      *scaledDim, op.getContext());
-
-  IndexExprsLatticeStorage newLattice =
-      IndexExprsLatticeStorage::join(toExprs[0], scaled);
-
-  if (newLattice.isTop() && !toExprs[0].isTop() && !scaled.isTop()) {
-    InFlightDiagnostic diag = emitError()
-                              << "conflict when propagating " << fromName
-                              << " to " << toName << " lattice in BitcastOp";
-    diag.attachNote() << toName << " lattice: " << toExprs[0];
-    diag.attachNote() << fromName << " lattice: " << fromExprs[0];
-    return diag;
-  }
-
-  return updateIfChanged(toExprs[0], newLattice);
-}
-
-llvm::FailureOr<ChangeResult> wave::BitcastOp::propagateIndexExprsForward(
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage> operandExprs,
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> resultExprs,
-    wave::EmitErrorFn emitError) {
-  WaveTensorType inputType =
-      llvm::dyn_cast<WaveTensorType>(getValueToCast().getType());
-  WaveTensorType resultType =
-      llvm::dyn_cast<WaveTensorType>(getResult().getType());
-  unsigned srcBits =
-      wave::getElementType(getValueToCast().getType()).getIntOrFloatBitWidth();
-  unsigned dstBits =
-      wave::getElementType(getResult().getType()).getIntOrFloatBitWidth();
-  return propagateBitcastIndexExprs(*this, operandExprs, resultExprs, inputType,
-                                    resultType, srcBits, dstBits, "operand",
-                                    "result", getResult(), emitError);
-}
-
-llvm::FailureOr<ChangeResult> wave::BitcastOp::propagateIndexExprsBackward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> operandExprs,
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage> resultExprs,
-    wave::EmitErrorFn emitError) {
-  WaveTensorType inputType =
-      llvm::dyn_cast<WaveTensorType>(getValueToCast().getType());
-  WaveTensorType resultType =
-      llvm::dyn_cast<WaveTensorType>(getResult().getType());
-  unsigned srcBits =
-      wave::getElementType(getValueToCast().getType()).getIntOrFloatBitWidth();
-  unsigned dstBits =
-      wave::getElementType(getResult().getType()).getIntOrFloatBitWidth();
-  return propagateBitcastIndexExprs(
-      *this, resultExprs, operandExprs, resultType, inputType, dstBits, srcBits,
-      "result", "operand", getValueToCast(), emitError);
 }
 
 llvm::FailureOr<mlir::ChangeResult>
@@ -3758,35 +2129,6 @@ llvm::SmallVector<WaveSymbolAttr> wave::BroadcastOp::inferBroadcastDims() {
   return broadcastDims;
 }
 
-FailureOr<ChangeResult> wave::BroadcastOp::propagateIndexExprsForward(
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage> operandIndexExprs,
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> resultIndexExprs,
-    wave::EmitErrorFn emitError) {
-  // Forward propagation is identity: it will propagate expressions for symbols
-  // present in the source to the result and make sure they are joined with
-  // those. Additional propagation backward from the result users will be needed
-  // to cover all symbols.
-  return detail::identityIndexExprsPropagate(operandIndexExprs,
-                                             resultIndexExprs, getResult(),
-                                             "operand", "result", emitError);
-}
-
-FailureOr<ChangeResult> wave::BroadcastOp::propagateIndexExprsBackward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> operandIndexExprs,
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage> resultIndexExprs,
-    wave::EmitErrorFn emitError) {
-  auto sourceTensorType = dyn_cast<WaveTensorType>(getSource().getType());
-  if (!sourceTensorType) {
-    emitError() << "expected source tensor type, got " << getSource().getType();
-    return failure();
-  }
-
-  // Backward propagation is identity only for symbols that are present
-  return detail::identityIndexExprsPropagate(
-      resultIndexExprs[0].keepOnlySymbols(sourceTensorType.getShape()),
-      operandIndexExprs, getSource(), "result", "operand", emitError);
-}
-
 LogicalResult wave::BroadcastOp::verify() {
   if (failed(detail::verifyElementTypesMatch(getLoc(), "source",
                                              getSource().getType(), "result",
@@ -3923,157 +2265,6 @@ llvm::FailureOr<ChangeResult> wave::PermuteOp::propagateBackward(
   // Cannot propagate shape information backward for permute operations
   // because the input shape ordering is not determined by the result.
   return ChangeResult::NoChange;
-}
-
-// Helper to permute strides in an index expressions lattice according to
-// the permutation from source shape to target shape.
-//
-// The permute operation swaps the strides of the permuted indices.
-// For example, if we have a permute operation that swaps [B, M, N] to
-// [M, N, B], then for each dimension k, we keep its start and step,
-// but take the stride from the dimension at the same position in
-// target_shape.
-static IndexExprsLatticeStorage
-permuteIndexExprsStrides(const IndexExprsLatticeStorage &inputLattice,
-                         llvm::ArrayRef<wave::WaveSymbolAttr> srcShape,
-                         llvm::ArrayRef<wave::WaveSymbolAttr> targetShape,
-                         MLIRContext *ctx, wave::EmitErrorFn emitError) {
-  if (inputLattice.isBottom() || inputLattice.isTop())
-    return inputLattice;
-
-  assert(srcShape.size() == targetShape.size() &&
-         "source shape rank does not match target shape rank");
-
-  DictionaryAttr inputDict = inputLattice.getConcreteValue();
-
-  llvm::DenseMap<WaveSymbolAttr, WaveIndexMappingAttr> symbolToMapping;
-  for (NamedAttribute namedAttr : inputDict) {
-    if (auto mapping =
-            llvm::dyn_cast<WaveIndexMappingAttr>(namedAttr.getValue())) {
-      auto key = WaveSymbolAttr::get(ctx, namedAttr.getName());
-      symbolToMapping[key] = mapping;
-    }
-  }
-
-  // Create the permuted index expressions.
-  // For each dimension k in src_shape:
-  //   - Keep start and step from the original mapping for k
-  //   - Take stride from the mapping for src_to_target[k]
-  SmallVector<NamedAttribute> permutedMappings;
-  permutedMappings.reserve(srcShape.size());
-  for (auto [srcSymbol, targetSymbol] :
-       llvm::zip_equal(srcShape, targetShape)) {
-    auto srcMappingIt = symbolToMapping.find(srcSymbol);
-    auto targetMappingIt = symbolToMapping.find(targetSymbol);
-
-    assert(srcMappingIt != symbolToMapping.end() &&
-           "source mapping not found for symbol");
-    assert(targetMappingIt != symbolToMapping.end() &&
-           "target mapping not found for symbol");
-
-    WaveIndexMappingAttr srcMapping = srcMappingIt->second;
-    WaveIndexMappingAttr targetMapping = targetMappingIt->second;
-
-    SmallVector<Attribute> allSymbols(srcMapping.getSymbols());
-    for (Attribute sym : targetMapping.getSymbols()) {
-      if (!llvm::is_contained(allSymbols, sym))
-        allSymbols.push_back(sym);
-    }
-
-    AffineMap alignedStart = alignMapSymbols(
-        srcMapping.getStart(), srcMapping.getSymbols(), allSymbols);
-    AffineMap alignedStep = alignMapSymbols(
-        srcMapping.getStep(), srcMapping.getSymbols(), allSymbols);
-    AffineMap alignedStride = alignMapSymbols(
-        targetMapping.getStride(), targetMapping.getSymbols(), allSymbols);
-
-    auto newMapping = WaveIndexMappingAttr::get(ctx, allSymbols, alignedStart,
-                                                alignedStep, alignedStride);
-
-    permutedMappings.push_back(
-        NamedAttribute(StringAttr::get(ctx, srcSymbol.getName()), newMapping));
-  }
-
-  return IndexExprsLatticeStorage(DictionaryAttr::get(ctx, permutedMappings),
-                                  inputLattice.getPriorities(),
-                                  inputLattice.getVectorShape())
-      .withSourceVectorShape(inputLattice.getSourceVectorShape(),
-                             inputLattice.getSourceVectorShapePriority());
-}
-
-llvm::FailureOr<ChangeResult> wave::PermuteOp::propagateIndexExprsForward(
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage> operandExprs,
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> resultExprs,
-    wave::EmitErrorFn emitError) {
-  auto inputType = llvm::dyn_cast<WaveTensorType>(getValue().getType());
-  if (!inputType || !inputType.getFullySpecified())
-    return ChangeResult::NoChange;
-
-  auto resultType = llvm::dyn_cast<WaveTensorType>(getResult().getType());
-  if (!resultType || !resultType.getFullySpecified())
-    return ChangeResult::NoChange;
-
-  ArrayRef<WaveSymbolAttr> targetShape = resultType.getShape();
-  ArrayRef<WaveSymbolAttr> srcShape = inputType.getShape();
-
-  IndexExprsLatticeStorage permuted = permuteIndexExprsStrides(
-      operandExprs[0], srcShape, targetShape, getContext(), emitError);
-
-  IndexExprsLatticeStorage newResultLattice =
-      IndexExprsLatticeStorage::join(resultExprs[0], permuted);
-
-  if (newResultLattice.isTop() && !resultExprs[0].isTop() &&
-      !permuted.isTop()) {
-    InFlightDiagnostic diag =
-        emitError()
-        << "conflict when propagating forward to the result lattice in "
-           "PermuteOp";
-    diag.attachNote() << "Result lattice: " << resultExprs[0];
-    diag.attachNote() << "Operand lattice: " << operandExprs[0];
-    return diag;
-  }
-
-  return updateIfChanged(resultExprs[0], newResultLattice);
-}
-
-llvm::FailureOr<ChangeResult> wave::PermuteOp::propagateIndexExprsBackward(
-    llvm::MutableArrayRef<wave::IndexExprsLatticeStorage> operandExprs,
-    llvm::ArrayRef<wave::IndexExprsLatticeStorage> resultExprs,
-    wave::EmitErrorFn emitError) {
-  auto inputType = llvm::dyn_cast<WaveTensorType>(getValue().getType());
-  if (!inputType || !inputType.getFullySpecified())
-    return ChangeResult::NoChange;
-
-  auto resultType = llvm::dyn_cast<WaveTensorType>(getResult().getType());
-  if (!resultType || !resultType.getFullySpecified())
-    return ChangeResult::NoChange;
-
-  if (!wave::detail::shouldPropagateIndexExprs(resultExprs[0], operandExprs[0],
-                                               getValue())) {
-    return ChangeResult::NoChange;
-  }
-
-  ArrayRef<WaveSymbolAttr> resultShape = resultType.getShape();
-  ArrayRef<WaveSymbolAttr> srcShape = inputType.getShape();
-
-  IndexExprsLatticeStorage permuted = permuteIndexExprsStrides(
-      resultExprs[0], resultShape, srcShape, getContext(), emitError);
-
-  IndexExprsLatticeStorage newOperandLattice =
-      IndexExprsLatticeStorage::join(operandExprs[0], permuted);
-
-  if (newOperandLattice.isTop() && !operandExprs[0].isTop() &&
-      !permuted.isTop()) {
-    InFlightDiagnostic diag =
-        emitError()
-        << "conflict when propagating backward to the operand lattice in "
-           "PermuteOp";
-    diag.attachNote() << "Operand lattice: " << operandExprs[0];
-    diag.attachNote() << "Result lattice: " << resultExprs[0];
-    return diag;
-  }
-
-  return updateIfChanged(operandExprs[0], newOperandLattice);
 }
 
 llvm::LogicalResult wave::PermuteOp::finalizeTypeInference() {
