+++ water/lib/Dialect/Wave/Transforms/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_mlir_dialect_library(MLIRWaveTransforms
   DetectNormalForms.cpp
   ExpandVariadicReductions.cpp
+  InferIndexExprs.cpp
   InferTypes.cpp
   LoweringPatterns.cpp
   LowerReadWriteOps.cpp
