+++ water/lib/Dialect/Wave/Transforms/InferTypes.cpp
@@ -35,12 +35,10 @@ using namespace mlir;
 #define DEBUG_TYPE "wave-infer-types"
 
 using wave::ElementsPerThreadLatticeValue;
-using wave::IndexExprsLatticeStorage;
 
 namespace wave {
 #define GEN_PASS_DEF_WATERWAVEINFERTYPESPASS
 #define GEN_PASS_DEF_WATERWAVEPROPAGATEELEMENTSPERTHREADPASS
-#define GEN_PASS_DEF_WATERWAVEINFERINDEXEXPRSPASS
 #include "water/Dialect/Wave/Transforms/Passes.h.inc"
 } // namespace wave
 
@@ -1286,899 +1284,3 @@ public:
   }
 };
 } // namespace
-
-// Lattice object for index expressions for analysis compatibility, the actual
-// logic is in the IndexExprsLatticeStorage class.
-class IndexExprsLattice : public dataflow::Lattice<IndexExprsLatticeStorage> {
-public:
-  MLIR_DEFINE_EXPLICIT_INTERNAL_INLINE_TYPE_ID(IndexExprsLattice);
-  using Lattice::Lattice;
-};
-
-class IndexExprsForwardAnalysis
-    : public dataflow::SparseForwardDataFlowAnalysis<IndexExprsLattice> {
-private:
-  void unsafeSet(IndexExprsLattice *lattice, IndexExprsLatticeStorage value) {
-    if (lattice->getValue() == value)
-      return;
-    lattice->getValue().unsafeSet(value);
-    propagateIfChanged(lattice, ChangeResult::Change);
-  }
-
-  void safeSet(IndexExprsLattice *lattice, IndexExprsLatticeStorage value) {
-    if (lattice->getValue() == value)
-      return;
-#ifndef NDEBUG
-    IndexExprsLatticeStorage joined =
-        IndexExprsLatticeStorage::join(lattice->getValue(), value);
-    assert(IndexExprsLatticeStorage::join(joined, lattice->getValue()) ==
-               joined &&
-           "join should not move the lattice backward, did you forget to join "
-           "with the original lattice value in an interface method "
-           "implementation?");
-    assert(
-        IndexExprsLatticeStorage::join(joined, value) == joined &&
-        "join should not move the lattice forward, did you forget to join with "
-        "the original lattice value in an interface method implementation?");
-#endif
-    unsafeSet(lattice, value);
-  }
-
-public:
-  explicit IndexExprsForwardAnalysis(
-      DataFlowSolver &solver,
-      wave::OverrideInitializationFn overrideInitialization = nullptr)
-      : SparseForwardDataFlowAnalysis(solver),
-        overrideInitialization(overrideInitialization) {}
-
-  LogicalResult initialize(Operation *top) override {
-    assert(!getSolverConfig().isInterprocedural() &&
-           "interprocedural analysis not supported");
-
-    // Call the base class initialization in order to set up update listeners.
-    // Note that this will initialize values at function/region entries to
-    // lattice top.
-    if (failed(SparseForwardDataFlowAnalysis::initialize(top)))
-      return failure();
-
-    llvm::DenseMap<Operation *, Attribute> constraints;
-    if (llvm::failed(wave::collectWaveConstraints(top, constraints)))
-      return llvm::failure();
-
-    for (auto &&[parent, attr] : constraints) {
-      wave::WaveHyperparameterAttr hyperparams =
-          wave::getHyperparameters(parent);
-      auto initObject =
-          wave::IndexExprsAnalysisInit::create(parent, attr, hyperparams);
-      if (llvm::failed(initObject))
-        return llvm::failure();
-      WalkResult walkResult = parent->walk([&](Operation *op) -> WalkResult {
-        if (auto iface =
-                llvm::dyn_cast<wave::WaveInferIndexExprsOpInterface>(op)) {
-          llvm::SmallVector<wave::IndexExprsLatticeStorage> resultExprs =
-              llvm::map_to_vector(op->getResults(), [&](Value v) {
-                return getLatticeElement(v)->getValue();
-              });
-          auto emitError = [op]() { return op->emitError(); };
-          LDBG() << "initializing index expressions forward for "
-                 << PrintNoRegions(op);
-          if (llvm::failed(iface.initializeIndexExprsForward(
-                  resultExprs, *initObject, emitError)))
-            return WalkResult::interrupt();
-
-          for (auto &&[result, lattice] :
-               llvm::zip_equal(op->getResults(), resultExprs)) {
-            IndexExprsLattice *latticeObject = getLatticeElement(result);
-            LDBG() << "  result #" << result.getResultNumber()
-                   << " original: " << *latticeObject;
-            safeSet(latticeObject, lattice);
-            LDBG() << "  result #" << result.getResultNumber()
-                   << " updated: " << *latticeObject;
-          }
-        }
-
-        if (auto iterateOp = llvm::dyn_cast<wave::IterateOp>(op)) {
-          // Set lattices of captured block arguments to the relevant tiling
-          // constraint, it will be then propagated by joining with
-          // expressions induced by other constraints.
-          wave::WaveSymbolAttr iterSymbolAttr = iterateOp.getIterator();
-          llvm::SmallVector<Attribute> symbolConstraints =
-              initObject->symbolConstraints.lookup(iterSymbolAttr);
-          auto it = llvm::find_if(symbolConstraints,
-                                  llvm::IsaPred<wave::TilingConstraintAttr>);
-          if (it != symbolConstraints.end()) {
-            wave::TilingConstraintAttr tilingConstraint =
-                llvm::cast<wave::TilingConstraintAttr>(*it);
-            for (Value capture : iterateOp.getCaptureBlockArgs()) {
-              auto captureType =
-                  dyn_cast<wave::WaveTensorType>(capture.getType());
-              if (!captureType)
-                continue;
-              if (!llvm::is_contained(captureType.getShape(), iterSymbolAttr))
-                continue;
-              auto dict = DictionaryAttr::get(
-                  iterSymbolAttr.getContext(),
-                  {{iterSymbolAttr.getName(),
-                    wave::applyConstraint(tilingConstraint)}});
-              LDBG() << "setting iterate block argument lattice " << capture
-                     << " from " << PrintNoRegions(iterateOp) << " to " << dict;
-              IndexExprsLattice *captureLattice = getLatticeElement(capture);
-              safeSet(
-                  captureLattice,
-                  wave::IndexExprsLatticeStorage::join(
-                      captureLattice->getValue(),
-                      wave::IndexExprsLatticeStorage(
-                          dict, wave::IndexExprsLatticeStorage::kLowestPriority,
-                          wave::detail::filterVectorShape(
-                              initObject->hardwareConstraint.getVectorShapes(),
-                              captureType.getShape()))));
-            }
-          }
-        }
-        return llvm::success();
-      });
-      if (walkResult.wasInterrupted())
-        return llvm::failure();
-    }
-
-    if (overrideInitialization) {
-      if (llvm::failed(overrideInitialization(
-              top, [&](Value value, DictionaryAttr dict,
-                       DictionaryAttr priorities, DictionaryAttr vecShape) {
-                if (!dict)
-                  return unsafeSet(getLatticeElement(value),
-                                   IndexExprsLatticeStorage::top());
-                unsafeSet(
-                    getLatticeElement(value),
-                    wave::IndexExprsLatticeStorage(dict, priorities, vecShape));
-              })))
-        return llvm::failure();
-    }
-
-    initialized = true;
-    return llvm::success();
-  }
-
-  void setToEntryState(IndexExprsLattice *lattice) override {
-    // Default logic will call this function on arguments of a callable
-    // operation since we are running in a non-interprocedural analysis. Setting
-    // them to top would propagate everywhere. Instead, just do nothing here and
-    // let them converge to whatever value needed by backward analysis.
-    auto arg = llvm::dyn_cast<BlockArgument>(lattice->getAnchor());
-    if (arg && llvm::isa<CallableOpInterface>(arg.getOwner()->getParentOp())) {
-      return;
-    }
-
-    // Default initialization calls `setToEntryState` on block arguments, we
-    // don't want to set it to the top state because it will propagate
-    // everywhere. Set/join with bottom instead so it can be overridden. Once
-    // initialization is done, `setToEntryState` may be called for unanalyzable
-    // cases, where we actually want to set it to the (pessimistic fixpoint) top
-    // state.
-    propagateIfChanged(lattice,
-                       lattice->join(initialized
-                                         ? IndexExprsLatticeStorage::top()
-                                         : IndexExprsLatticeStorage::bottom()));
-    if (initialized) {
-      LDBG() << "top fixpoint for " << lattice->getAnchor() << " "
-             << (arg ? PrintNoRegions(arg.getOwner()->getParentOp())
-                     : PrintNoRegions(nullptr));
-    }
-  }
-
-  llvm::LogicalResult
-  visitOperation(Operation *op,
-                 llvm::ArrayRef<const IndexExprsLattice *> operands,
-                 llvm::ArrayRef<IndexExprsLattice *> results) override {
-
-    LLVM_DEBUG({
-      LDBG() << "visiting operation forward " << PrintNoRegions(op);
-      LDBG() << "  operand lattices:";
-      for (auto [i, operand] : llvm::enumerate(operands)) {
-        LDBG() << "    operand #" << i << ": " << *operand;
-      }
-      // Print all result lattices.
-      LDBG() << "  result lattices:";
-      for (auto [i, result] : llvm::enumerate(results)) {
-        LDBG() << "    result #" << i << ": " << *result;
-      }
-    });
-    llvm::scope_exit scope([&] {
-      LLVM_DEBUG({
-        LDBG() << "  updated result lattices:";
-        for (auto [i, result] : llvm::enumerate(results)) {
-          LDBG() << "    result #" << i << ": " << *result;
-        }
-      });
-    });
-
-    // Check if the operation implements the interface.
-    if (!llvm::isa<wave::WaveInferIndexExprsOpInterface>(op)) {
-      // Operations without the interface should not manipulate WaveTensorType.
-      if (!llvm::any_of(op->getOperandTypes(),
-                        llvm::IsaPred<wave::WaveTensorType>) &&
-          !llvm::any_of(op->getResultTypes(),
-                        llvm::IsaPred<wave::WaveTensorType>)) {
-        return llvm::success();
-      }
-      return op->emitError()
-             << "cannot propagate index expressions across an operation not "
-                "implementing the wave infer index expressions interface";
-    }
-
-    auto extractLattice = [](const IndexExprsLattice *lattice) {
-      return lattice->getValue();
-    };
-    llvm::SmallVector<IndexExprsLatticeStorage> operandLattices =
-        llvm::map_to_vector(operands, extractLattice);
-    llvm::SmallVector<IndexExprsLatticeStorage> resultLattices =
-        llvm::map_to_vector(results, extractLattice);
-
-    auto reportError = [op]() { return op->emitError(); };
-    llvm::FailureOr<ChangeResult> result =
-        llvm::cast<wave::WaveInferIndexExprsOpInterface>(op)
-            .propagateIndexExprsForward(operandLattices, resultLattices,
-                                        reportError);
-    if (llvm::failed(result))
-      return llvm::failure();
-    if (*result == ChangeResult::NoChange)
-      return llvm::success();
-
-    for (auto &&[resultLattice, lattice] :
-         llvm::zip_equal(resultLattices, results)) {
-      safeSet(lattice, resultLattice);
-    }
-    return llvm::success();
-  }
-
-  void visitNonControlFlowArguments(
-      Operation *op, const RegionSuccessor &successor,
-      ValueRange nonSuccessorInputs,
-      llvm::ArrayRef<IndexExprsLattice *> lattices) override {
-    auto iterateOp = llvm::dyn_cast<wave::IterateOp>(op);
-    if (!iterateOp)
-      return;
-
-    // Technically, the non-captured arguments can be seen as forwarded from
-    // operands or results, but they need special handling to remove
-    // loop-specific parts of the index.
-    assert((successor.isParent() ||
-            successor.getSuccessor()->getRegionNumber() == 0) &&
-           "unexpected control flow");
-
-    auto yieldOp =
-        llvm::cast<wave::YieldOp>(iterateOp.getLoopBody()->getTerminator());
-
-    LDBG() << "visiting " << PrintNoRegions(iterateOp);
-    if (successor.getSuccessor()) {
-      LDBG() << " propagating to region #"
-             << successor.getSuccessor()->getRegionNumber();
-
-      // When successor is the body region, propagate induction variable
-      // lattices from their initial and yielded values.
-      for (auto &&[terminatorOperand, iterArg, lattice] : llvm::zip_equal(
-               yieldOp.getOperands(), iterateOp.getIterArgs(),
-               lattices.take_front(iterateOp.getIterArgs().size()))) {
-        // See comments in
-        // InferTypeForwardAnalysis::visitNonControlFlowArguments.
-        const IndexExprsLattice *iterArgLattice = getLatticeElementFor(
-            getProgramPointBefore(iterateOp.getLoopBody()), iterArg);
-        const IndexExprsLattice *terminatorOperandLattice =
-            getLatticeElementFor(getProgramPointBefore(iterateOp.getLoopBody()),
-                                 terminatorOperand);
-        LDBG() << "iter arg lattice: " << *iterArgLattice;
-        LDBG() << "terminator operand lattice: " << *terminatorOperandLattice;
-        LDBG() << "block lattice: " << *lattice;
-        ChangeResult changed = lattice->join(iterArgLattice->getValue());
-        changed |= lattice->join(terminatorOperandLattice->getValue());
-        propagateIfChanged(lattice, changed);
-        LDBG() << "new block lattice: " << *lattice;
-      }
-
-      // And also propagate lattices for captured values, which only need to be
-      // propagated from the initial values.
-      for (auto &&[capture, lattice] : llvm::zip_equal(
-               iterateOp.getCaptures(),
-               lattices.take_back(iterateOp.getCaptures().size()))) {
-        // See comments in
-        // InferTypeForwardAnalysis::visitNonControlFlowArguments.
-        const IndexExprsLattice *captureLattice = getLatticeElementFor(
-            getProgramPointBefore(iterateOp.getLoopBody()), capture);
-
-        LDBG() << "captured lattice: " << *captureLattice;
-        LDBG() << "block lattice: " << *lattice;
-        propagateIfChanged(lattice, lattice->join(captureLattice->getValue()));
-        LDBG() << "new block lattice: " << *lattice;
-      }
-    } else {
-      LDBG() << "propagating to parent";
-
-      // When successor is the iterate op itself, propagate lattices from iter
-      // args and terminator operands to results while removing loop-specific
-      // parts of the index.
-      for (auto &&[terminatorOperand, iterArg, resultLattice] : llvm::zip_equal(
-               yieldOp.getOperands(), iterateOp.getIterArgs(), lattices)) {
-        // See comments in
-        // InferTypeForwardAnalysis::visitNonControlFlowArguments.
-        const IndexExprsLattice *terminatorOperandLattice =
-            getLatticeElementFor(getProgramPointAfter(iterateOp),
-                                 terminatorOperand);
-        const IndexExprsLattice *iterArgLattice =
-            getLatticeElementFor(getProgramPointAfter(iterateOp), iterArg);
-        LDBG() << "iter arg lattice: " << *iterArgLattice;
-        LDBG() << "terminator operand lattice: " << *terminatorOperandLattice;
-        LDBG() << "result lattice: " << *resultLattice;
-        ChangeResult changed = resultLattice->join(iterArgLattice->getValue());
-        changed |= resultLattice->join(
-            terminatorOperandLattice->getValue().withoutIterSymbols(
-                iterateOp.getIterator()));
-        propagateIfChanged(resultLattice, changed);
-        LDBG() << "new result lattice: " << *resultLattice;
-      }
-    }
-  }
-
-  // Return true if there are pending error reports.
-  bool hasDelayedErrors() const { return !delayedErrors.empty(); }
-
-  // Return the emitter of a pending error report for the given operation.
-  wave::EmitDelayedErrorFn getDelayedError(Operation *op) const {
-    return delayedErrors.lookup_or(op, wave::EmitDelayedErrorFn());
-  }
-
-private:
-  bool initialized = false;
-  wave::OverrideInitializationFn overrideInitialization;
-  llvm::SmallDenseMap<Operation *, wave::EmitDelayedErrorFn> delayedErrors;
-};
-
-class IndexExprsBackwardAnalysis
-    : public dataflow::SparseBackwardDataFlowAnalysis<IndexExprsLattice> {
-private:
-  void unsafeSet(IndexExprsLattice *lattice, IndexExprsLatticeStorage value) {
-    if (lattice->getValue() == value)
-      return;
-    lattice->getValue().unsafeSet(value);
-    propagateIfChanged(lattice, ChangeResult::Change);
-  }
-
-  void safeSet(IndexExprsLattice *lattice, IndexExprsLatticeStorage value) {
-    if (lattice->getValue() == value)
-      return;
-#ifndef NDEBUG
-    IndexExprsLatticeStorage joined =
-        IndexExprsLatticeStorage::join(lattice->getValue(), value);
-    assert(IndexExprsLatticeStorage::join(joined, lattice->getValue()) ==
-               joined &&
-           "join should not move the lattice backward, did you forget to join "
-           "with the original lattice value in an interface method "
-           "implementation?");
-    assert(
-        IndexExprsLatticeStorage::join(joined, value) == joined &&
-        "join should not move the lattice forward, did you forget to join with "
-        "the original lattice value in an interface method implementation?");
-#endif
-    unsafeSet(lattice, value);
-  }
-
-public:
-  IndexExprsBackwardAnalysis(
-      DataFlowSolver &solver, SymbolTableCollection &symbolTable,
-      wave::OverrideInitializationFn overrideInitialization = nullptr)
-      : SparseBackwardDataFlowAnalysis(solver, symbolTable),
-        overrideInitialization(overrideInitialization) {}
-
-  llvm::LogicalResult initialize(Operation *top) override {
-    assert(!getSolverConfig().isInterprocedural() &&
-           "interprocedural analysis not supported");
-
-    // Call the base class initialization in order to set up update listeners.
-    // Note that this will initialize values at function/region entries to
-    // lattice top.
-    if (llvm::failed(SparseBackwardDataFlowAnalysis::initialize(top)))
-      return llvm::failure();
-
-    llvm::DenseMap<Operation *, Attribute> constraints;
-    if (llvm::failed(wave::collectWaveConstraints(top, constraints)))
-      return llvm::failure();
-    for (auto &&[parent, attr] : constraints) {
-      wave::WaveHyperparameterAttr hyperparams =
-          wave::getHyperparameters(parent);
-      auto initObject =
-          wave::IndexExprsAnalysisInit::create(parent, attr, hyperparams);
-      if (llvm::failed(initObject))
-        return llvm::failure();
-
-      WalkResult walkResult = parent->walk([&](Operation *op) -> WalkResult {
-        if (op->hasTrait<wave::RequiresSidewaysBackwardPropagationOpTrait>()) {
-          for (Value operand : op->getOperands())
-            addDependency(getLatticeElement(operand), getProgramPointAfter(op));
-        }
-        if (auto iface =
-                llvm::dyn_cast<wave::WaveInferIndexExprsOpInterface>(op)) {
-          llvm::SmallVector<wave::IndexExprsLatticeStorage> operandExprs =
-              llvm::map_to_vector(op->getOperands(), [&](Value v) {
-                return getLatticeElement(v)->getValue();
-              });
-          auto emitError = [op]() { return op->emitError(); };
-
-          LDBG() << "initializing index expressions backward for "
-                 << PrintNoRegions(op);
-          wave::EmitDelayedErrorFn delayedErrorEmitter = nullptr;
-          if (llvm::failed(iface.initializeIndexExprsBackward(
-                  operandExprs, *initObject, emitError, delayedErrorEmitter)))
-            return WalkResult::interrupt();
-          if (delayedErrorEmitter) {
-            LDBG() << "delayed error recorded\n";
-            delayedErrors[op] = delayedErrorEmitter;
-          }
-          for (auto &&[i, operand, lattice] :
-               llvm::enumerate(op->getOperands(), operandExprs)) {
-            IndexExprsLattice *latticeObject = getLatticeElement(operand);
-            LDBG() << "  operand #" << i << " original: " << *latticeObject;
-            safeSet(latticeObject, lattice);
-            LDBG() << "  operand #" << i << " updated: " << *latticeObject;
-          }
-          return WalkResult::advance();
-        }
-
-        return WalkResult::advance();
-      });
-      if (walkResult.wasInterrupted())
-        return failure();
-    }
-
-    if (overrideInitialization) {
-      if (llvm::failed(overrideInitialization(
-              top, [&](Value value, DictionaryAttr dict,
-                       DictionaryAttr priorities, DictionaryAttr vecShape) {
-                if (!dict)
-                  return unsafeSet(getLatticeElement(value),
-                                   IndexExprsLatticeStorage::top());
-                unsafeSet(
-                    getLatticeElement(value),
-                    wave::IndexExprsLatticeStorage(dict, priorities, vecShape));
-              })))
-        return llvm::failure();
-    }
-
-    initialized = true;
-    return llvm::success();
-  }
-
-  void visitBranchOperand(OpOperand &opOperand) override {
-    if (!llvm::isa<wave::WaveTensorType>(opOperand.get().getType()))
-      return;
-
-    // Captures of the iterate need to be propagated from the corresponding
-    // block arguments manually without the tiling constraint.
-    if (auto iterateOp =
-            llvm::dyn_cast<wave::IterateOp>(opOperand.getOwner())) {
-      unsigned position = opOperand.getOperandNumber();
-      Value blockArgument = iterateOp.getLoopBody()->getArgument(position);
-      IndexExprsLattice *blockArgLattice = getLatticeElement(blockArgument);
-      IndexExprsLattice *lattice = getLatticeElement(opOperand.get());
-      // Explicitly add a dependency to update this analysis at the program
-      // point before the iterate if the block argument lattice changes. Using
-      // localized const cast here to avoid accidentally modifying this lattice
-      // in the function.
-      // TODO: expose it upstream and use.
-      addDependency(blockArgLattice, getProgramPointAfter(iterateOp));
-
-      if (!wave::detail::shouldPropagateIndexExprs(blockArgLattice->getValue(),
-                                                   lattice->getValue(),
-                                                   opOperand.get())) {
-        return;
-      }
-
-      LDBG() << "propagating backwards from block argument #" << position
-             << " to op operand " << PrintNoRegions(iterateOp);
-      LDBG() << "block argument lattice: " << *blockArgLattice;
-      LDBG() << "lattice: " << *lattice;
-      IndexExprsLatticeStorage joined = IndexExprsLatticeStorage::join(
-          lattice->getValue(), blockArgLattice->getValue().withoutIterSymbols(
-                                   iterateOp.getIterator()));
-      safeSet(lattice, joined);
-      LDBG() << "new lattice: " << *lattice;
-      return;
-    }
-
-    // Terminator operands are propagated from op results as is.
-    if (auto yieldOp = llvm::dyn_cast<wave::YieldOp>(opOperand.getOwner())) {
-      unsigned position = opOperand.getOperandNumber();
-      Value result = yieldOp->getParentOp()->getResult(position);
-      const IndexExprsLattice *resultLattice = getLatticeElement(result);
-      IndexExprsLattice *lattice = getLatticeElement(opOperand.get());
-      // Explicitly add a dependency to update this analysis at the program
-      // point before the terminator if the result lattice changes. Using
-      // localized const cast here to avoid accidentally modifying this lattice.
-      // TODO: expose it upstream and use.
-      addDependency(const_cast<IndexExprsLattice *>(resultLattice),
-                    getProgramPointAfter(yieldOp));
-
-      if (!wave::detail::shouldPropagateIndexExprs(resultLattice->getValue(),
-                                                   lattice->getValue(),
-                                                   opOperand.get())) {
-        return;
-      }
-
-      LDBG() << "propagating backwards from region-carrying op result #"
-             << position << " to terminator operand " << yieldOp;
-      LDBG() << "result lattice: " << *resultLattice;
-      LDBG() << "lattice: " << *lattice;
-      IndexExprsLatticeStorage joined = IndexExprsLatticeStorage::join(
-          lattice->getValue(), resultLattice->getValue());
-      safeSet(lattice, joined);
-      LDBG() << "new lattice: " << *lattice;
-      return;
-    }
-
-    setToExitState(getLatticeElement(opOperand.get()));
-  }
-
-  void visitCallOperand(OpOperand &opOperand) override {
-    if (!llvm::isa<wave::WaveTensorType>(opOperand.get().getType()))
-      return;
-    setToExitState(getLatticeElement(opOperand.get()));
-  }
-
-  void setToExitState(IndexExprsLattice *lattice) override {
-    // Default initialization calls `setToExitState` on terminator and call
-    // operands, we don't want to set it to the top state because it will
-    // propagate everywhere. Set/join with bottom instead so it can be
-    // overridden. Once initialization is done, `setToExitState` may be called
-    // for unanalyzable cases, where we actually want to set it to the
-    // (pessimistic fixpoint) top state.
-    propagateIfChanged(lattice,
-                       lattice->join(initialized
-                                         ? IndexExprsLatticeStorage::top()
-                                         : IndexExprsLatticeStorage::bottom()));
-    if (initialized)
-      LDBG() << "top fixpoint (backward) for " << lattice->getAnchor();
-  }
-
-  llvm::LogicalResult
-  visitOperation(Operation *op, llvm::ArrayRef<IndexExprsLattice *> operands,
-                 llvm::ArrayRef<const IndexExprsLattice *> results) override {
-    LLVM_DEBUG({
-      LDBG() << "visiting operation backward " << PrintNoRegions(op);
-      LDBG() << "  operand lattices:";
-      for (auto [i, operand] : llvm::enumerate(operands)) {
-        LDBG() << "    operand #" << i << ": " << *operand;
-      }
-      LDBG() << "  results lattices:";
-      for (auto [i, result] : llvm::enumerate(results)) {
-        LDBG() << "    result #" << i << ": " << *result;
-      }
-    });
-    llvm::scope_exit scope([&] {
-      LLVM_DEBUG({
-        LDBG() << "  updated operand lattices:";
-        for (auto [i, operand] : llvm::enumerate(operands)) {
-          LDBG() << "    operand #" << i << ": " << *operand;
-        }
-      });
-    });
-
-    // Check if the operation implements the interface.
-    if (!llvm::isa<wave::WaveInferIndexExprsOpInterface>(op)) {
-      // Operations without the interface should not manipulate WaveTensorType.
-      if (!llvm::any_of(op->getOperandTypes(),
-                        llvm::IsaPred<wave::WaveTensorType>) &&
-          !llvm::any_of(op->getResultTypes(),
-                        llvm::IsaPred<wave::WaveTensorType>)) {
-        return llvm::success();
-      }
-      return op->emitError()
-             << "cannot propagate index expressions across an operation not "
-                "implementing the wave infer index expressions interface";
-    }
-
-    auto extractLattice = [](const IndexExprsLattice *lattice) {
-      return lattice->getValue();
-    };
-    llvm::SmallVector<IndexExprsLatticeStorage> operandLattices =
-        llvm::map_to_vector(operands, extractLattice);
-    llvm::SmallVector<IndexExprsLatticeStorage> resultLattices =
-        llvm::map_to_vector(results, extractLattice);
-
-    auto reportError = [op]() { return op->emitError(); };
-    llvm::FailureOr<mlir::ChangeResult> result =
-        llvm::cast<wave::WaveInferIndexExprsOpInterface>(op)
-            .propagateIndexExprsBackward(operandLattices, resultLattices,
-                                         reportError);
-    if (llvm::failed(result))
-      return llvm::failure();
-    if (*result == ChangeResult::NoChange)
-      return llvm::success();
-
-    for (auto &&[operandLattice, lattice] :
-         llvm::zip_equal(operandLattices, operands)) {
-      safeSet(lattice, operandLattice);
-    }
-    return llvm::success();
-  }
-
-  // Visit the non-forwarded arguments of a region, such as the
-  // induction variables of a loop.
-  void
-  visitNonControlFlowArguments(RegionSuccessor & /*successor*/,
-                               ArrayRef<BlockArgument> /*arguments*/) override {
-    // This is called for induction variables of an IterateOp, which is handled
-    // by the forward analysis.
-  }
-
-  // Returns true if there are any delayed errors.
-  bool hasDelayedErrors() const { return !delayedErrors.empty(); }
-
-  // Returns the delayed error emitter for the given operation.
-  wave::EmitDelayedErrorFn getDelayedError(Operation *op) const {
-    return delayedErrors.lookup_or(op, wave::EmitDelayedErrorFn());
-  }
-
-private:
-  bool initialized = false;
-  wave::OverrideInitializationFn overrideInitialization;
-  llvm::SmallDenseMap<Operation *, wave::EmitDelayedErrorFn, 4> delayedErrors;
-};
-
-namespace {
-class InferIndexExprsPass
-    : public wave::impl::WaterWaveInferIndexExprsPassBase<InferIndexExprsPass> {
-public:
-  using Base::Base;
-
-  void runOnOperation() override {
-    if (llvm::failed(wave::verifyNormalFormPassPrecondition(
-            {wave::WaveNormalForm::FunctionBoundarySpecified,
-             wave::WaveNormalForm::OpTypesSpecified},
-            getOperation(), getArgument())))
-      return signalPassFailure();
-
-    IRRewriter rewriter(&getContext());
-    getOperation()->walk(
-        [&](wave::IterateOp iterateOp) { iterateOp.makeIsolated(rewriter); });
-
-    SymbolTableCollection symbolTable;
-    DataFlowConfig config;
-    config.setInterprocedural(false);
-    DataFlowSolver solver(config);
-
-    solver.load<mlir::dataflow::DeadCodeAnalysis>();
-    solver.load<mlir::dataflow::SparseConstantPropagation>();
-    wave::DelayedErrorEmitterInfo delayedErrorInfo =
-        wave::addWaveIndexExprsAnalyses(solver, symbolTable);
-
-    if (llvm::failed(
-            wave::runSolverAndCaptureErrors(solver, getOperation(), false)))
-      return signalPassFailure();
-
-    if (llvm::failed(wave::setWaveIndexExprAnalysisResults(
-            getOperation(), solver, delayedErrorInfo)))
-      return signalPassFailure();
-
-    getOperation()->walk([&](wave::IterateOp iterateOp) {
-      iterateOp.makeNonIsolated(rewriter);
-    });
-
-    if (llvm::failed(wave::setNormalFormPassPostcondition(
-            {wave::WaveNormalForm::IndexExprsSpecified}, getOperation())))
-      return signalPassFailure();
-  }
-};
-} // namespace
-
-wave::DelayedErrorEmitterInfo
-wave::addWaveIndexExprsAnalyses(DataFlowSolver &solver,
-                                SymbolTableCollection &symbolTable,
-                                wave::WaveIndexExprsAnalysisOptions options) {
-  IndexExprsForwardAnalysis *forward = nullptr;
-  if (!options.disableForward) {
-    forward =
-        solver.load<IndexExprsForwardAnalysis>(options.overrideInitialization);
-  }
-  IndexExprsBackwardAnalysis *backward = nullptr;
-  if (!options.disableBackward) {
-    backward = solver.load<IndexExprsBackwardAnalysis>(
-        symbolTable, options.overrideInitialization);
-  }
-
-  // Note that these lambdas are stored and used later so they must not capture
-  // anything that has a function-level lifetime.
-  wave::DelayedErrorEmitterInfo delayedErrorEmitterInfo;
-  delayedErrorEmitterInfo.getDelayedError =
-      [forward, backward](Operation *op) -> wave::EmitDelayedErrorFn {
-    if (forward) {
-      if (wave::EmitDelayedErrorFn delayedError = forward->getDelayedError(op))
-        return delayedError;
-    }
-    if (backward) {
-      return backward->getDelayedError(op);
-    }
-    return nullptr;
-  };
-  delayedErrorEmitterInfo.hasDelayedErrors = [forward, backward]() {
-    return (forward && forward->hasDelayedErrors()) ||
-           (backward && backward->hasDelayedErrors());
-  };
-  return delayedErrorEmitterInfo;
-}
-
-/// If any slot has a vector shape, every slot must have one; otherwise emit an
-/// error. Empty shape dictionaries are not used as placeholders on the IR.
-static LogicalResult collectPerValueVectorShapeAttrs(
-    Operation *op, llvm::ArrayRef<wave::IndexExprsLatticeStorage> slots,
-    llvm::function_ref<void(llvm::raw_ostream &, unsigned)> describeSlot,
-    llvm::SmallVectorImpl<Attribute> &shapeDicts) {
-  bool anyPresent =
-      llvm::any_of(slots, [](const wave::IndexExprsLatticeStorage &l) {
-        return l.getVectorShape() != nullptr;
-      });
-  if (!anyPresent)
-    return success();
-  shapeDicts.reserve(shapeDicts.size() + slots.size());
-  for (const auto [i, lat] : llvm::enumerate(slots)) {
-    DictionaryAttr vs = lat.getVectorShape();
-    if (!vs) {
-      llvm::SmallString<64> buf;
-      llvm::raw_svector_ostream os(buf);
-      describeSlot(os, i);
-      return op->emitError() << "missing vector shape for " << buf.str();
-    }
-    shapeDicts.push_back(vs);
-  }
-  return success();
-}
-
-LogicalResult wave::setWaveIndexExprAnalysisResults(
-    Operation *top, const DataFlowSolver &solver,
-    const DelayedErrorEmitterInfo &delayedErrorInfo,
-    function_ref<LogicalResult(Operation *, ArrayRef<IndexExprsLatticeStorage>)>
-        extraHandler) {
-  bool hadFailures = false;
-  WalkResult walkResult =
-      top->walk([&](wave::WaveInferIndexExprsOpInterface iface) {
-        auto getLatticeValue = [&](Value value) {
-          auto *latticeObject = solver.lookupState<IndexExprsLattice>(value);
-          return latticeObject ? latticeObject->getValue()
-                               : IndexExprsLatticeStorage::bottom();
-        };
-
-        SmallVector<Value> valuesForIndexExpr;
-        SmallVector<Attribute> indexExprs;
-
-        // Special case for MMA operations. We always set their index attribute
-        // to whatever is implied by the MMA kind regardless of the values
-        // inferred for operands because MMAs must retain their specific index
-        // expressions.
-        // TODO: this shouldn't strictly necessary in a purely MLIR flow and is
-        // kept for Python compatibility.
-        std::function<void(raw_ostream &, unsigned)> descriptionGenerator =
-            iface.getIndexExprValuesAndDescriptions(valuesForIndexExpr);
-        indexExprs.reserve(valuesForIndexExpr.size());
-
-        // Shared handler for MMA-family ops: initializes operand lattices from
-        // MMA-kind-implied index expressions and sets the result index
-        // attribute. `numOperands` is the number of MMA operands (3 for MmaOp,
-        // 5 for ScaledMmaOp). `resultDescription` is used in error messages.
-        auto handleMmaOp = [&](wave::WaveInferIndexExprsOpInterface mmaIface,
-                               unsigned numOperands,
-                               StringRef resultDescription) -> WalkResult {
-          Operation *mmaOp = mmaIface.getOperation();
-          SmallVector<wave::IndexExprsLatticeStorage> operandLattices(
-              numOperands, wave::IndexExprsLatticeStorage::bottom());
-          wave::EmitDelayedErrorFn delayedError;
-          Attribute constraints;
-          // TODO(#1049): this is not ideal, especially when combined with
-          // getting hyperparameters below, we could just have a double walk
-          // with a kernel operation first if we had one, or even do a
-          // per-kernel pass.
-          for (Operation *parent = mmaOp->getParentOp(); parent && !constraints;
-               parent = parent->getParentOp()) {
-            constraints = parent->getAttrOfType<ArrayAttr>(
-                wave::WaveDialect::kWaveConstraintsAttrName);
-          }
-          if (!constraints) {
-            mmaOp->emitError("constraints not found");
-            return WalkResult::interrupt();
-          }
-          FailureOr<wave::IndexExprsAnalysisInit> init =
-              wave::IndexExprsAnalysisInit::create(
-                  mmaOp, constraints, wave::getHyperparameters(mmaOp));
-          if (failed(init))
-            return WalkResult::interrupt();
-          if (failed(mmaIface.initializeIndexExprsBackward(
-                  operandLattices, *init, [&]() { return mmaOp->emitError(); },
-                  delayedError))) {
-            if (delayedError) {
-              InFlightDiagnostic diag = mmaOp->emitError();
-              delayedError(diag);
-            }
-            return WalkResult::interrupt();
-          }
-
-          for (auto &&[i, lattice] : llvm::enumerate(operandLattices)) {
-            SmallString<32> description;
-            llvm::raw_svector_ostream os(description);
-            descriptionGenerator(os, i);
-            [[maybe_unused]] LogicalResult logicalResult =
-                detail::checkAndAppendIndexExpr(mmaOp->getLoc(), lattice,
-                                                description, indexExprs);
-            assert(succeeded(logicalResult) &&
-                   "failed to append implied index expression, it must not be "
-                   "bottom/top");
-          }
-          Value result = mmaOp->getResult(0);
-          if (failed(detail::checkAndAppendIndexExpr(
-                  mmaOp->getLoc(), getLatticeValue(result), resultDescription,
-                  indexExprs)))
-            return WalkResult::interrupt();
-
-          MLIRContext *ctx = mmaIface->getContext();
-          SmallVector<wave::IndexExprsLatticeStorage> slots(operandLattices);
-          slots.push_back(getLatticeValue(result));
-          SmallVector<Attribute> shapeDicts;
-          if (failed(collectPerValueVectorShapeAttrs(
-                  mmaOp, slots, descriptionGenerator, shapeDicts)))
-            return WalkResult::interrupt();
-          mmaIface->setAttr(wave::WaveDialect::kIndexWaveExprListAttrName,
-                            ArrayAttr::get(ctx, indexExprs));
-          if (!shapeDicts.empty())
-            mmaIface->setAttr(wave::WaveDialect::kVectorShapeAttrName,
-                              ArrayAttr::get(ctx, shapeDicts));
-          return WalkResult::advance();
-        };
-
-        if (auto mma = dyn_cast<wave::MmaOp>(iface.getOperation()))
-          return handleMmaOp(mma, /*numOperands=*/3, "mma result");
-
-        if (auto scaledMma = dyn_cast<wave::ScaledMmaOp>(iface.getOperation()))
-          return handleMmaOp(scaledMma, /*numOperands=*/5, "scaled_mma result");
-
-        for (auto &&[i, value] : llvm::enumerate(valuesForIndexExpr)) {
-          llvm::SmallString<32> description;
-          llvm::raw_svector_ostream os(description);
-          descriptionGenerator(os, i);
-          if (failed(detail::checkAndAppendIndexExpr(iface->getLoc(),
-                                                     getLatticeValue(value),
-                                                     os.str(), indexExprs))) {
-            // Don't stop on the first reported error if there are some delayed
-            // errors that would be useful to report here. We need to wait and
-            // see whether the operation they are attached to actually has had
-            // inference issues as some errors may be corrected.
-            if (!delayedErrorInfo.hasDelayedErrors())
-              return WalkResult::interrupt();
-
-            hadFailures = true;
-            if (auto delayedError = delayedErrorInfo.getDelayedError(iface)) {
-              InFlightDiagnostic diag =
-                  iface->emitError()
-                  << "the error above may be caused by the following: ";
-              delayedError(diag);
-            }
-          }
-        }
-
-        SmallVector<wave::IndexExprsLatticeStorage> slots = llvm::map_to_vector(
-            valuesForIndexExpr, [&](Value v) { return getLatticeValue(v); });
-        // Only set the index expressions if there were no failures.
-        if (!hadFailures) {
-          MLIRContext *ctx = iface->getContext();
-          SmallVector<Attribute> shapeDicts;
-          if (failed(collectPerValueVectorShapeAttrs(
-                  iface, slots, descriptionGenerator, shapeDicts)))
-            return WalkResult::interrupt();
-          iface->setAttr(wave::WaveDialect::kIndexWaveExprListAttrName,
-                         ArrayAttr::get(ctx, indexExprs));
-          if (!shapeDicts.empty())
-            iface->setAttr(wave::WaveDialect::kVectorShapeAttrName,
-                           ArrayAttr::get(ctx, shapeDicts));
-        }
-
-        if (extraHandler && failed(extraHandler(iface, slots)))
-          return WalkResult::interrupt();
-
-        return WalkResult::advance();
-      });
-  return llvm::failure(hadFailures || walkResult.wasInterrupted());
-}
